export opcode data to content folder + export SVG to public folder + nits
