+++ public/opCodeSVG.tsx
@@ -1,0 +1,277 @@
+import { motion } from "framer-motion"
+import React from "react"
+
+interface OpcodeSVGProps extends React.ComponentPropsWithoutRef<"svg"> {
+    svgRef?: React.RefObject<SVGSVGElement>
+}
+
+export function OpcodeSVG({ svgRef, ...props }: OpcodeSVGProps) {
+    return (
+        <svg
+            ref={svgRef}
+            width="120%"
+            height="auto"
+            viewBox="0 0 1063 545"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+            {...props}
+        >
+            <rect
+                x="292.5"
+                y="147.5"
+                width="338"
+                height="53"
+                fill="#D9D9D9"
+                stroke="black"
+            />
+            <path
+                d="M425.985 182C425.579 182 425.223 181.848 424.918 181.543C424.614 181.238 424.461 180.883 424.461 180.476V162.19C424.461 161.784 424.614 161.429 424.918 161.124C425.223 160.819 425.579 160.667 425.985 160.667H435.128C436.347 160.667 437.414 161.124 438.328 162.038C439.242 162.952 439.699 164.019 439.699 165.238V177.429C439.699 178.648 439.242 179.714 438.328 180.629C437.414 181.543 436.347 182 435.128 182H425.985ZM436.652 177.429V165.238C436.652 164.832 436.499 164.476 436.195 164.171C435.89 163.867 435.534 163.714 435.128 163.714H427.509V178.952H435.128C435.534 178.952 435.89 178.8 436.195 178.495C436.499 178.19 436.652 177.835 436.652 177.429ZM447.314 182C446.095 182 445.028 181.543 444.114 180.629C443.2 179.714 442.743 178.648 442.743 177.429C442.743 176.21 443.2 175.143 444.114 174.229C445.028 173.314 446.095 172.857 447.314 172.857H454.933V171.333C454.933 170.927 454.781 170.571 454.476 170.267C454.171 169.962 453.816 169.81 453.409 169.81H447.314C446.908 169.81 446.552 169.657 446.247 169.352C445.943 169.048 445.79 168.692 445.79 168.286C445.79 167.879 445.943 167.524 446.247 167.219C446.552 166.914 446.908 166.762 447.314 166.762H453.409C454.628 166.762 455.695 167.219 456.609 168.133C457.524 169.048 457.981 170.114 457.981 171.333V180.476C457.981 180.883 457.828 181.238 457.524 181.543C457.219 181.848 456.863 182 456.457 182H447.314ZM454.933 175.905H447.314C446.908 175.905 446.552 176.057 446.247 176.362C445.943 176.667 445.79 177.022 445.79 177.429C445.79 177.835 445.943 178.19 446.247 178.495C446.552 178.8 446.908 178.952 447.314 178.952H454.933V175.905ZM465.595 169.81C465.189 169.81 464.833 169.657 464.529 169.352C464.224 169.048 464.071 168.692 464.071 168.286C464.071 167.879 464.224 167.524 464.529 167.219C464.833 166.914 465.189 166.762 465.595 166.762H467.119V162.19C467.119 161.784 467.271 161.429 467.576 161.124C467.881 160.819 468.237 160.667 468.643 160.667C469.049 160.667 469.405 160.819 469.71 161.124C470.014 161.429 470.167 161.784 470.167 162.19V166.762H474.738C475.144 166.762 475.5 166.914 475.805 167.219C476.11 167.524 476.262 167.879 476.262 168.286C476.262 168.692 476.11 169.048 475.805 169.352C475.5 169.657 475.144 169.81 474.738 169.81H470.167V177.429C470.167 177.835 470.319 178.19 470.624 178.495C470.929 178.8 471.284 178.952 471.69 178.952H474.738C475.144 178.952 475.5 179.105 475.805 179.41C476.11 179.714 476.262 180.07 476.262 180.476C476.262 180.883 476.11 181.238 475.805 181.543C475.5 181.848 475.144 182 474.738 182H471.69C470.471 182 469.405 181.543 468.49 180.629C467.576 179.714 467.119 178.648 467.119 177.429V169.81H465.595ZM483.876 182C482.657 182 481.591 181.543 480.676 180.629C479.762 179.714 479.305 178.648 479.305 177.429C479.305 176.21 479.762 175.143 480.676 174.229C481.591 173.314 482.657 172.857 483.876 172.857H491.496V171.333C491.496 170.927 491.343 170.571 491.038 170.267C490.734 169.962 490.378 169.81 489.972 169.81H483.876C483.47 169.81 483.115 169.657 482.81 169.352C482.505 169.048 482.353 168.692 482.353 168.286C482.353 167.879 482.505 167.524 482.81 167.219C483.115 166.914 483.47 166.762 483.876 166.762H489.972C491.191 166.762 492.257 167.219 493.172 168.133C494.086 169.048 494.543 170.114 494.543 171.333V180.476C494.543 180.883 494.391 181.238 494.086 181.543C493.781 181.848 493.426 182 493.019 182H483.876ZM491.496 175.905H483.876C483.47 175.905 483.115 176.057 482.81 176.362C482.505 176.667 482.353 177.022 482.353 177.429C482.353 177.835 482.505 178.19 482.81 178.495C483.115 178.8 483.47 178.952 483.876 178.952H491.496V175.905Z"
+                fill="black"
+            />
+            <path
+                d="M448.41 112.273H449.672L452.638 119.517H452.74L455.706 112.273H456.967V121H455.979V114.369H455.893L453.166 121H452.211L449.484 114.369H449.399V121H448.41V112.273ZM460.876 121.153C460.461 121.153 460.085 121.075 459.747 120.919C459.409 120.76 459.14 120.531 458.941 120.233C458.742 119.932 458.643 119.568 458.643 119.142C458.643 118.767 458.717 118.463 458.865 118.23C459.012 117.994 459.21 117.81 459.457 117.676C459.704 117.543 459.977 117.443 460.275 117.378C460.576 117.31 460.879 117.256 461.183 117.216C461.58 117.165 461.903 117.126 462.15 117.101C462.4 117.072 462.582 117.026 462.695 116.96C462.812 116.895 462.87 116.781 462.87 116.619V116.585C462.87 116.165 462.755 115.838 462.525 115.605C462.298 115.372 461.953 115.256 461.49 115.256C461.009 115.256 460.633 115.361 460.36 115.571C460.088 115.781 459.896 116.006 459.785 116.244L458.83 115.903C459.001 115.506 459.228 115.196 459.512 114.974C459.799 114.75 460.112 114.594 460.45 114.506C460.791 114.415 461.126 114.369 461.455 114.369C461.666 114.369 461.907 114.395 462.18 114.446C462.455 114.494 462.721 114.595 462.977 114.749C463.235 114.902 463.45 115.134 463.62 115.443C463.791 115.753 463.876 116.168 463.876 116.688V121H462.87V120.114H462.819C462.751 120.256 462.637 120.408 462.478 120.57C462.319 120.732 462.107 120.869 461.843 120.983C461.579 121.097 461.257 121.153 460.876 121.153ZM461.029 120.25C461.427 120.25 461.762 120.172 462.035 120.016C462.311 119.859 462.518 119.658 462.657 119.411C462.799 119.163 462.87 118.903 462.87 118.631V117.71C462.828 117.761 462.734 117.808 462.589 117.851C462.447 117.891 462.282 117.926 462.095 117.957C461.91 117.986 461.73 118.011 461.553 118.034C461.38 118.054 461.24 118.071 461.132 118.085C460.87 118.119 460.626 118.175 460.399 118.251C460.174 118.325 459.992 118.437 459.853 118.588C459.717 118.736 459.649 118.937 459.649 119.193C459.649 119.543 459.778 119.807 460.036 119.986C460.298 120.162 460.629 120.25 461.029 120.25ZM466.461 114.455L468.03 117.131L469.598 114.455H470.757L468.643 117.727L470.757 121H469.598L468.03 118.46L466.461 121H465.302L467.382 117.727L465.302 114.455H466.461ZM436.189 136.119C435.689 136.119 435.239 136.02 434.838 135.821C434.438 135.622 434.117 135.349 433.875 135.003C433.634 134.656 433.502 134.261 433.479 133.818H434.502C434.542 134.213 434.721 134.54 435.039 134.798C435.36 135.054 435.743 135.182 436.189 135.182C436.547 135.182 436.865 135.098 437.144 134.93C437.425 134.763 437.645 134.533 437.804 134.24C437.966 133.945 438.047 133.611 438.047 133.239C438.047 132.858 437.963 132.518 437.796 132.22C437.631 131.919 437.404 131.682 437.114 131.509C436.824 131.335 436.493 131.247 436.121 131.244C435.854 131.241 435.58 131.283 435.299 131.368C435.017 131.45 434.786 131.557 434.604 131.688L433.615 131.568L434.144 127.273H438.678V128.21H435.03L434.723 130.784H434.775C434.953 130.642 435.178 130.524 435.448 130.43C435.718 130.337 435.999 130.29 436.292 130.29C436.826 130.29 437.301 130.418 437.719 130.673C438.14 130.926 438.469 131.273 438.708 131.713C438.949 132.153 439.07 132.656 439.07 133.222C439.07 133.778 438.945 134.276 438.695 134.713C438.448 135.148 438.107 135.491 437.672 135.744C437.238 135.994 436.743 136.119 436.189 136.119ZM440.763 136V135.233L443.643 132.08C443.982 131.71 444.26 131.389 444.479 131.116C444.697 130.841 444.859 130.582 444.964 130.341C445.072 130.097 445.126 129.841 445.126 129.574C445.126 129.267 445.053 129.001 444.905 128.777C444.76 128.553 444.561 128.379 444.308 128.257C444.055 128.135 443.771 128.074 443.456 128.074C443.121 128.074 442.828 128.143 442.578 128.283C442.331 128.419 442.139 128.611 442.003 128.858C441.869 129.105 441.803 129.395 441.803 129.727H440.797C440.797 129.216 440.915 128.767 441.151 128.381C441.386 127.994 441.707 127.693 442.114 127.477C442.523 127.261 442.982 127.153 443.49 127.153C444.001 127.153 444.455 127.261 444.849 127.477C445.244 127.693 445.554 127.984 445.778 128.351C446.003 128.717 446.115 129.125 446.115 129.574C446.115 129.895 446.057 130.209 445.94 130.516C445.827 130.82 445.628 131.159 445.344 131.534C445.062 131.906 444.672 132.361 444.172 132.898L442.212 134.994V135.062H446.268V136H440.763ZM450.875 136.119C450.233 136.119 449.686 135.945 449.234 135.595C448.783 135.243 448.438 134.733 448.199 134.065C447.96 133.395 447.841 132.585 447.841 131.636C447.841 130.693 447.96 129.888 448.199 129.22C448.44 128.55 448.787 128.038 449.239 127.686C449.693 127.331 450.239 127.153 450.875 127.153C451.511 127.153 452.055 127.331 452.507 127.686C452.962 128.038 453.308 128.55 453.547 129.22C453.788 129.888 453.909 130.693 453.909 131.636C453.909 132.585 453.79 133.395 453.551 134.065C453.312 134.733 452.967 135.243 452.516 135.595C452.064 135.945 451.517 136.119 450.875 136.119ZM450.875 135.182C451.511 135.182 452.006 134.875 452.358 134.261C452.71 133.648 452.886 132.773 452.886 131.636C452.886 130.881 452.805 130.237 452.643 129.706C452.484 129.175 452.254 128.77 451.953 128.491C451.655 128.213 451.295 128.074 450.875 128.074C450.244 128.074 449.751 128.385 449.396 129.007C449.041 129.626 448.864 130.503 448.864 131.636C448.864 132.392 448.943 133.034 449.102 133.562C449.261 134.091 449.49 134.493 449.788 134.768C450.089 135.044 450.452 135.182 450.875 135.182ZM459.057 136V127.273H460.062V130.494H460.148C460.222 130.381 460.324 130.236 460.455 130.06C460.588 129.881 460.778 129.722 461.026 129.582C461.276 129.44 461.614 129.369 462.04 129.369C462.591 129.369 463.077 129.507 463.497 129.783C463.918 130.058 464.246 130.449 464.482 130.955C464.717 131.46 464.835 132.057 464.835 132.744C464.835 133.438 464.717 134.038 464.482 134.547C464.246 135.053 463.919 135.445 463.501 135.723C463.084 135.999 462.602 136.136 462.057 136.136C461.636 136.136 461.3 136.067 461.047 135.928C460.794 135.786 460.599 135.625 460.463 135.446C460.327 135.264 460.222 135.114 460.148 134.994H460.028V136H459.057ZM460.045 132.727C460.045 133.222 460.118 133.658 460.263 134.036C460.408 134.411 460.619 134.705 460.898 134.918C461.176 135.128 461.517 135.233 461.92 135.233C462.341 135.233 462.692 135.122 462.973 134.901C463.257 134.676 463.47 134.375 463.612 133.997C463.757 133.616 463.83 133.193 463.83 132.727C463.83 132.267 463.759 131.852 463.616 131.483C463.477 131.111 463.266 130.817 462.982 130.601C462.7 130.382 462.347 130.273 461.92 130.273C461.511 130.273 461.168 130.376 460.889 130.584C460.611 130.788 460.401 131.075 460.259 131.445C460.116 131.811 460.045 132.239 460.045 132.727ZM466.787 138.455C466.616 138.455 466.464 138.44 466.331 138.412C466.197 138.386 466.105 138.361 466.054 138.335L466.31 137.449C466.554 137.511 466.77 137.534 466.957 137.517C467.145 137.5 467.311 137.416 467.456 137.266C467.604 137.118 467.739 136.878 467.861 136.545L468.048 136.034L465.628 129.455H466.719L468.526 134.67H468.594L470.401 129.455H471.491L468.713 136.955C468.588 137.293 468.433 137.572 468.249 137.794C468.064 138.018 467.849 138.185 467.605 138.293C467.364 138.401 467.091 138.455 466.787 138.455ZM475.665 129.455V130.307H472.273V129.455H475.665ZM473.262 127.886H474.268V134.125C474.268 134.409 474.309 134.622 474.391 134.764C474.477 134.903 474.585 134.997 474.715 135.045C474.849 135.091 474.989 135.114 475.137 135.114C475.248 135.114 475.339 135.108 475.41 135.097C475.481 135.082 475.538 135.071 475.58 135.062L475.785 135.966C475.717 135.991 475.621 136.017 475.499 136.043C475.377 136.071 475.222 136.085 475.035 136.085C474.751 136.085 474.472 136.024 474.2 135.902C473.93 135.78 473.705 135.594 473.526 135.344C473.35 135.094 473.262 134.778 473.262 134.398V127.886ZM479.852 136.136C479.222 136.136 478.678 135.997 478.22 135.719C477.766 135.438 477.415 135.045 477.168 134.543C476.923 134.037 476.801 133.449 476.801 132.778C476.801 132.108 476.923 131.517 477.168 131.006C477.415 130.491 477.759 130.091 478.199 129.804C478.642 129.514 479.159 129.369 479.75 129.369C480.091 129.369 480.428 129.426 480.76 129.54C481.092 129.653 481.395 129.838 481.668 130.094C481.94 130.347 482.158 130.682 482.32 131.099C482.482 131.517 482.562 132.031 482.562 132.642V133.068H477.517V132.199H481.54C481.54 131.83 481.466 131.5 481.318 131.21C481.173 130.92 480.966 130.692 480.696 130.524C480.429 130.357 480.114 130.273 479.75 130.273C479.349 130.273 479.003 130.372 478.71 130.571C478.42 130.767 478.197 131.023 478.041 131.338C477.885 131.653 477.807 131.991 477.807 132.352V132.932C477.807 133.426 477.892 133.845 478.062 134.189C478.236 134.53 478.476 134.79 478.783 134.969C479.089 135.145 479.446 135.233 479.852 135.233C480.116 135.233 480.355 135.196 480.568 135.122C480.784 135.045 480.97 134.932 481.126 134.781C481.283 134.628 481.403 134.437 481.489 134.21L482.46 134.483C482.358 134.812 482.186 135.102 481.945 135.352C481.703 135.599 481.405 135.793 481.05 135.932C480.695 136.068 480.295 136.136 479.852 136.136ZM488.729 130.92L487.825 131.176C487.768 131.026 487.685 130.879 487.574 130.737C487.466 130.592 487.318 130.473 487.131 130.379C486.943 130.286 486.703 130.239 486.411 130.239C486.01 130.239 485.676 130.331 485.409 130.516C485.145 130.697 485.013 130.929 485.013 131.21C485.013 131.46 485.104 131.658 485.286 131.803C485.467 131.947 485.751 132.068 486.138 132.165L487.109 132.403C487.695 132.545 488.131 132.763 488.418 133.055C488.705 133.345 488.848 133.719 488.848 134.176C488.848 134.551 488.74 134.886 488.524 135.182C488.311 135.477 488.013 135.71 487.629 135.881C487.246 136.051 486.8 136.136 486.291 136.136C485.624 136.136 485.071 135.991 484.634 135.702C484.196 135.412 483.919 134.989 483.803 134.432L484.757 134.193C484.848 134.545 485.02 134.81 485.273 134.986C485.528 135.162 485.862 135.25 486.274 135.25C486.743 135.25 487.115 135.151 487.391 134.952C487.669 134.75 487.808 134.509 487.808 134.227C487.808 134 487.729 133.81 487.57 133.656C487.411 133.5 487.166 133.384 486.837 133.307L485.746 133.051C485.146 132.909 484.706 132.689 484.425 132.391C484.146 132.089 484.007 131.713 484.007 131.261C484.007 130.892 484.111 130.565 484.318 130.281C484.528 129.997 484.814 129.774 485.175 129.612C485.538 129.45 485.95 129.369 486.411 129.369C487.058 129.369 487.567 129.511 487.936 129.795C488.308 130.08 488.572 130.455 488.729 130.92Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M92.2707 54.3153L92.5008 52.9091H102.319L102.089 54.3153H97.9724L96.0292 66H94.444L96.3872 54.3153H92.2707ZM104.868 60.0938L103.896 66H102.388L104.561 52.9091H106.07L105.277 57.7159H105.405C105.712 57.2045 106.121 56.7997 106.632 56.5014C107.148 56.2031 107.795 56.054 108.575 56.054C109.24 56.054 109.8 56.1882 110.256 56.4567C110.717 56.7209 111.04 57.1278 111.228 57.6776C111.42 58.223 111.447 58.9176 111.311 59.7614L110.263 66H108.754L109.777 59.8636C109.905 59.0838 109.8 58.4808 109.464 58.0547C109.131 57.6243 108.614 57.4091 107.91 57.4091C107.416 57.4091 106.956 57.5135 106.53 57.7223C106.108 57.9311 105.75 58.2358 105.456 58.6364C105.166 59.0369 104.97 59.5227 104.868 60.0938ZM113.022 66L114.659 56.1818H116.167L114.531 66H113.022ZM115.873 54.5199C115.579 54.5199 115.326 54.4197 115.113 54.2195C114.904 54.0192 114.799 53.7784 114.799 53.4972C114.799 53.2159 114.904 52.9751 115.113 52.7749C115.326 52.5746 115.579 52.4744 115.873 52.4744C116.167 52.4744 116.419 52.5746 116.628 52.7749C116.841 52.9751 116.947 53.2159 116.947 53.4972C116.947 53.7784 116.841 54.0192 116.628 54.2195C116.419 54.4197 116.167 54.5199 115.873 54.5199ZM125.527 58.3807L124.095 58.7642C124.044 58.5341 123.954 58.3125 123.827 58.0994C123.703 57.8821 123.511 57.7053 123.251 57.5689C122.996 57.4283 122.646 57.358 122.203 57.358C121.602 57.358 121.08 57.4964 120.637 57.7734C120.194 58.0462 119.94 58.3935 119.876 58.8153C119.812 59.1861 119.898 59.4822 120.132 59.7038C120.366 59.9254 120.758 60.1065 121.308 60.2472L122.714 60.6051C123.558 60.8224 124.159 61.1484 124.517 61.5831C124.879 62.0135 125.003 62.5739 124.888 63.2642C124.794 63.8267 124.549 64.3295 124.153 64.7727C123.756 65.2159 123.249 65.5653 122.631 65.821C122.013 66.0767 121.325 66.2045 120.567 66.2045C119.557 66.2045 118.76 65.9872 118.176 65.5526C117.592 65.1136 117.281 64.4787 117.243 63.6477L118.751 63.2898C118.79 63.8139 118.977 64.2102 119.314 64.4787C119.655 64.7429 120.132 64.875 120.746 64.875C121.449 64.875 122.035 64.7259 122.503 64.4276C122.972 64.125 123.239 63.7628 123.302 63.3409C123.366 62.9957 123.298 62.7081 123.098 62.478C122.902 62.2479 122.561 62.0753 122.075 61.9602L120.516 61.5767C119.646 61.3594 119.037 61.027 118.687 60.5795C118.342 60.1321 118.227 59.5696 118.342 58.892C118.436 58.3381 118.675 57.848 119.058 57.4219C119.446 56.9957 119.932 56.6612 120.516 56.4183C121.104 56.1754 121.743 56.054 122.433 56.054C123.405 56.054 124.131 56.267 124.613 56.6932C125.094 57.1151 125.399 57.6776 125.527 58.3807ZM135.455 66.2045C134.633 66.2045 133.943 65.9979 133.384 65.5845C132.826 65.1669 132.432 64.5788 132.202 63.8203C131.976 63.0575 131.948 62.1562 132.119 61.1165C132.293 60.0852 132.619 59.1903 133.097 58.4318C133.578 57.6733 134.168 57.0874 134.867 56.674C135.566 56.2607 136.329 56.054 137.156 56.054C137.795 56.054 138.283 56.1605 138.619 56.3736C138.96 56.5824 139.207 56.821 139.361 57.0895C139.514 57.3537 139.631 57.571 139.712 57.7415H139.84L140.646 52.9091H142.154L139.981 66H138.524L138.779 64.4915H138.6C138.46 64.6705 138.264 64.8963 138.012 65.169C137.761 65.4375 137.428 65.6783 137.015 65.8913C136.606 66.1001 136.086 66.2045 135.455 66.2045ZM135.826 64.8494C136.431 64.8494 136.972 64.6918 137.45 64.3764C137.931 64.0568 138.332 63.6158 138.651 63.0533C138.971 62.4865 139.193 61.8324 139.316 61.0909C139.44 60.358 139.435 59.7166 139.303 59.1669C139.171 58.6129 138.918 58.1825 138.543 57.8757C138.168 57.5646 137.671 57.4091 137.053 57.4091C136.418 57.4091 135.867 57.5732 135.398 57.9013C134.929 58.2251 134.548 58.6662 134.254 59.2244C133.96 59.7784 133.755 60.4006 133.64 61.0909C133.525 61.7898 133.523 62.4247 133.634 62.9957C133.744 63.5625 133.979 64.0142 134.337 64.3509C134.699 64.6832 135.195 64.8494 135.826 64.8494ZM145.804 66.2301C145.182 66.2301 144.636 66.1129 144.168 65.8786C143.699 65.6399 143.352 65.2969 143.126 64.8494C142.9 64.3977 142.838 63.8523 142.94 63.2131C143.038 62.6506 143.228 62.1946 143.509 61.8452C143.791 61.4915 144.131 61.2145 144.532 61.0142C144.937 60.8097 145.372 60.6584 145.836 60.5604C146.305 60.4581 146.771 60.3793 147.236 60.3239C147.841 60.2472 148.333 60.1875 148.712 60.1449C149.092 60.1023 149.377 60.032 149.569 59.9339C149.761 59.8359 149.878 59.6676 149.921 59.429V59.3778C150.023 58.7514 149.94 58.2635 149.671 57.9141C149.403 57.5604 148.923 57.3835 148.233 57.3835C147.517 57.3835 146.918 57.5412 146.437 57.8565C145.955 58.1719 145.6 58.5085 145.369 58.8665L144.04 58.3551C144.389 57.7585 144.807 57.294 145.293 56.9616C145.778 56.625 146.288 56.3906 146.82 56.2585C147.353 56.1222 147.867 56.054 148.361 56.054C148.676 56.054 149.032 56.0923 149.428 56.169C149.825 56.2415 150.195 56.3928 150.541 56.6229C150.89 56.853 151.152 57.2003 151.327 57.6648C151.506 58.1293 151.531 58.7514 151.403 59.5312L150.33 66H148.821L149.051 64.6705H148.975C148.838 64.8835 148.629 65.1115 148.348 65.3544C148.071 65.5973 147.72 65.804 147.293 65.9744C146.872 66.1449 146.375 66.2301 145.804 66.2301ZM146.239 64.875C146.835 64.875 147.357 64.7578 147.805 64.5234C148.256 64.2891 148.619 63.9865 148.891 63.6158C149.168 63.245 149.341 62.8551 149.409 62.446L149.639 61.0653C149.563 61.142 149.411 61.2124 149.185 61.2763C148.96 61.3359 148.704 61.3892 148.418 61.4361C148.137 61.4787 147.862 61.517 147.594 61.5511C147.33 61.581 147.117 61.6065 146.955 61.6278C146.554 61.679 146.173 61.7621 145.81 61.8771C145.453 61.9879 145.15 62.1562 144.903 62.3821C144.66 62.6037 144.509 62.9062 144.449 63.2898C144.359 63.8139 144.485 64.2102 144.826 64.4787C145.167 64.7429 145.638 64.875 146.239 64.875ZM158.989 56.1818L158.785 57.4602H153.697L153.901 56.1818H158.989ZM155.768 53.8295H157.276L155.717 63.1875C155.648 63.6136 155.659 63.9332 155.748 64.1463C155.842 64.3551 155.983 64.4957 156.17 64.5682C156.362 64.6364 156.569 64.6705 156.79 64.6705C156.957 64.6705 157.093 64.6619 157.199 64.6449C157.306 64.6236 157.391 64.6065 157.455 64.5938L157.557 65.9489C157.442 65.9872 157.287 66.0256 157.091 66.0639C156.899 66.1065 156.663 66.1278 156.381 66.1278C155.955 66.1278 155.555 66.0362 155.18 65.853C154.805 65.6697 154.517 65.3906 154.317 65.0156C154.116 64.6406 154.063 64.1676 154.157 63.5966L155.768 53.8295ZM162.486 66.2301C161.864 66.2301 161.318 66.1129 160.849 65.8786C160.381 65.6399 160.033 65.2969 159.807 64.8494C159.582 64.3977 159.52 63.8523 159.622 63.2131C159.72 62.6506 159.91 62.1946 160.191 61.8452C160.472 61.4915 160.813 61.2145 161.214 61.0142C161.619 60.8097 162.053 60.6584 162.518 60.5604C162.986 60.4581 163.453 60.3793 163.918 60.3239C164.523 60.2472 165.015 60.1875 165.394 60.1449C165.773 60.1023 166.059 60.032 166.251 59.9339C166.442 59.8359 166.56 59.6676 166.602 59.429V59.3778C166.704 58.7514 166.621 58.2635 166.353 57.9141C166.084 57.5604 165.605 57.3835 164.915 57.3835C164.199 57.3835 163.6 57.5412 163.119 57.8565C162.637 58.1719 162.281 58.5085 162.051 58.8665L160.722 58.3551C161.071 57.7585 161.489 57.294 161.974 56.9616C162.46 56.625 162.969 56.3906 163.502 56.2585C164.035 56.1222 164.548 56.054 165.043 56.054C165.358 56.054 165.714 56.0923 166.11 56.169C166.506 56.2415 166.877 56.3928 167.222 56.6229C167.572 56.853 167.834 57.2003 168.008 57.6648C168.187 58.1293 168.213 58.7514 168.085 59.5312L167.011 66H165.503L165.733 64.6705H165.656C165.52 64.8835 165.311 65.1115 165.03 65.3544C164.753 65.5973 164.401 65.804 163.975 65.9744C163.553 66.1449 163.057 66.2301 162.486 66.2301ZM162.92 64.875C163.517 64.875 164.039 64.7578 164.486 64.5234C164.938 64.2891 165.3 63.9865 165.573 63.6158C165.85 63.245 166.023 62.8551 166.091 62.446L166.321 61.0653C166.244 61.142 166.093 61.2124 165.867 61.2763C165.641 61.3359 165.386 61.3892 165.1 61.4361C164.819 61.4787 164.544 61.517 164.275 61.5511C164.011 61.581 163.798 61.6065 163.636 61.6278C163.236 61.679 162.854 61.7621 162.492 61.8771C162.134 61.9879 161.832 62.1562 161.584 62.3821C161.342 62.6037 161.19 62.9062 161.131 63.2898C161.041 63.8139 161.167 64.2102 161.508 64.4787C161.849 64.7429 162.32 64.875 162.92 64.875ZM178.79 66.2045C177.866 66.2045 177.107 65.9872 176.515 65.5526C175.927 65.1179 175.517 64.5192 175.287 63.7564C175.057 62.9936 175.023 62.1222 175.185 61.142C175.347 60.1449 175.675 59.2649 176.169 58.5021C176.664 57.7351 177.28 57.1364 178.017 56.706C178.754 56.2713 179.566 56.054 180.452 56.054C181.142 56.054 181.743 56.1818 182.255 56.4375C182.766 56.6932 183.16 57.0511 183.437 57.5114C183.714 57.9716 183.844 58.5085 183.827 59.1222H182.319C182.276 58.6747 182.084 58.2784 181.743 57.9332C181.407 57.5838 180.908 57.4091 180.248 57.4091C179.66 57.4091 179.118 57.5625 178.624 57.8693C178.13 58.1719 177.714 58.6001 177.378 59.1541C177.041 59.7038 176.813 60.3494 176.694 61.0909C176.566 61.8494 176.576 62.5099 176.726 63.0724C176.875 63.6349 177.145 64.0717 177.537 64.3828C177.929 64.6939 178.424 64.8494 179.02 64.8494C179.412 64.8494 179.779 64.7812 180.12 64.6449C180.461 64.5085 180.763 64.3125 181.027 64.0568C181.292 63.8011 181.5 63.4943 181.654 63.1364H183.162C182.975 63.7159 182.674 64.2379 182.261 64.7024C181.848 65.1626 181.347 65.5291 180.759 65.8018C180.171 66.0703 179.515 66.2045 178.79 66.2045ZM188.934 66.2045C188.031 66.2045 187.275 65.9893 186.665 65.5589C186.056 65.1286 185.625 64.5256 185.374 63.75C185.127 62.9744 185.086 62.071 185.253 61.0398C185.415 60.0341 185.736 59.1584 186.218 58.4126C186.704 57.6626 187.307 57.0831 188.027 56.674C188.751 56.2607 189.548 56.054 190.417 56.054C191.321 56.054 192.075 56.2713 192.68 56.706C193.29 57.1364 193.72 57.7415 193.971 58.5213C194.223 59.2969 194.265 60.2045 194.099 61.2443C193.937 62.2415 193.611 63.1129 193.121 63.8587C192.635 64.6001 192.032 65.1776 191.312 65.5909C190.596 66 189.804 66.2045 188.934 66.2045ZM188.96 64.8494C189.65 64.8494 190.249 64.6726 190.756 64.3189C191.267 63.9652 191.681 63.4986 191.996 62.919C192.316 62.3395 192.531 61.7131 192.642 61.0398C192.744 60.4006 192.74 59.8061 192.629 59.2564C192.518 58.7024 192.282 58.2571 191.919 57.9205C191.557 57.5795 191.048 57.4091 190.392 57.4091C189.701 57.4091 189.098 57.5881 188.583 57.946C188.071 58.304 187.658 58.7749 187.343 59.3587C187.027 59.9425 186.816 60.571 186.71 61.2443C186.608 61.8835 186.61 62.478 186.716 63.0277C186.827 63.5732 187.064 64.0142 187.426 64.3509C187.792 64.6832 188.304 64.8494 188.96 64.8494ZM202.474 61.9858L203.446 56.1818H204.954L203.318 66H201.81L202.091 64.3381H201.989C201.682 64.8366 201.253 65.2607 200.704 65.6101C200.154 65.9553 199.508 66.1278 198.767 66.1278C198.153 66.1278 197.629 65.9936 197.195 65.7251C196.764 65.4524 196.457 65.0433 196.274 64.4979C196.095 63.9482 196.074 63.2557 196.21 62.4205L197.258 56.1818H198.767L197.744 62.3182C197.629 63.0341 197.738 63.6051 198.07 64.0312C198.403 64.4574 198.882 64.6705 199.508 64.6705C199.883 64.6705 200.28 64.5746 200.697 64.3828C201.119 64.1911 201.494 63.897 201.822 63.5007C202.155 63.1044 202.372 62.5994 202.474 61.9858ZM209.763 52.9091L207.59 66H206.081L208.254 52.9091H209.763ZM214.047 66.2045C213.225 66.2045 212.534 65.9979 211.976 65.5845C211.418 65.1669 211.024 64.5788 210.794 63.8203C210.568 63.0575 210.54 62.1562 210.71 61.1165C210.885 60.0852 211.211 59.1903 211.688 58.4318C212.17 57.6733 212.76 57.0874 213.459 56.674C214.158 56.2607 214.921 56.054 215.747 56.054C216.387 56.054 216.875 56.1605 217.211 56.3736C217.552 56.5824 217.799 56.821 217.953 57.0895C218.106 57.3537 218.223 57.571 218.304 57.7415H218.432L219.237 52.9091H220.746L218.573 66H217.115L217.371 64.4915H217.192C217.051 64.6705 216.855 64.8963 216.604 65.169C216.353 65.4375 216.02 65.6783 215.607 65.8913C215.198 66.1001 214.678 66.2045 214.047 66.2045ZM214.418 64.8494C215.023 64.8494 215.564 64.6918 216.041 64.3764C216.523 64.0568 216.924 63.6158 217.243 63.0533C217.563 62.4865 217.784 61.8324 217.908 61.0909C218.032 60.358 218.027 59.7166 217.895 59.1669C217.763 58.6129 217.509 58.1825 217.134 57.8757C216.759 57.5646 216.263 57.4091 215.645 57.4091C215.01 57.4091 214.458 57.5732 213.99 57.9013C213.521 58.2251 213.139 58.6662 212.845 59.2244C212.551 59.7784 212.347 60.4006 212.232 61.0909C212.117 61.7898 212.115 62.4247 212.225 62.9957C212.336 63.5625 212.571 64.0142 212.929 64.3509C213.291 64.6832 213.787 64.8494 214.418 64.8494ZM226.799 66L228.973 52.9091H230.481L229.688 57.7415H229.816C229.953 57.571 230.142 57.3537 230.385 57.0895C230.628 56.821 230.954 56.5824 231.363 56.3736C231.772 56.1605 232.296 56.054 232.936 56.054C233.762 56.054 234.455 56.2607 235.013 56.674C235.576 57.0874 235.97 57.6733 236.196 58.4318C236.426 59.1903 236.456 60.0852 236.285 61.1165C236.115 62.1562 235.789 63.0575 235.307 63.8203C234.83 64.5788 234.242 65.1669 233.543 65.5845C232.848 65.9979 232.092 66.2045 231.274 66.2045C230.643 66.2045 230.155 66.1001 229.81 65.8913C229.469 65.6783 229.218 65.4375 229.056 65.169C228.898 64.8963 228.777 64.6705 228.691 64.4915H228.512L228.257 66H226.799ZM229.1 61.0909C228.981 61.8324 228.983 62.4865 229.107 63.0533C229.23 63.6158 229.475 64.0568 229.842 64.3764C230.208 64.6918 230.694 64.8494 231.299 64.8494C231.93 64.8494 232.484 64.6832 232.961 64.3509C233.438 64.0142 233.831 63.5625 234.137 62.9957C234.444 62.4247 234.657 61.7898 234.777 61.0909C234.887 60.4006 234.883 59.7784 234.764 59.2244C234.649 58.6662 234.406 58.2251 234.035 57.9013C233.669 57.5732 233.166 57.4091 232.527 57.4091C231.913 57.4091 231.37 57.5646 230.897 57.8757C230.428 58.1825 230.04 58.6129 229.733 59.1669C229.431 59.7166 229.22 60.358 229.1 61.0909ZM241.865 66.2045C240.924 66.2045 240.146 65.9957 239.532 65.5781C238.919 65.1562 238.49 64.5682 238.247 63.8139C238.005 63.0554 237.966 62.1733 238.132 61.1676C238.299 60.1619 238.629 59.2756 239.123 58.5085C239.617 57.7372 240.233 57.1364 240.97 56.706C241.708 56.2713 242.517 56.054 243.399 56.054C243.911 56.054 244.403 56.1392 244.876 56.3097C245.349 56.4801 245.758 56.7571 246.103 57.1406C246.448 57.5199 246.689 58.0227 246.826 58.6491C246.962 59.2756 246.953 60.0469 246.8 60.9631L246.698 61.6023H239.13L239.334 60.2983H245.368C245.466 59.7443 245.441 59.25 245.291 58.8153C245.147 58.3807 244.893 58.0376 244.531 57.7862C244.173 57.5348 243.719 57.4091 243.169 57.4091C242.603 57.4091 242.074 57.5625 241.584 57.8693C241.094 58.1719 240.685 58.5554 240.357 59.0199C240.029 59.4801 239.824 59.9489 239.743 60.4261L239.564 61.5511C239.436 62.3651 239.456 63.0128 239.622 63.4943C239.792 63.9759 240.088 64.3232 240.51 64.5362C240.932 64.745 241.46 64.8494 242.095 64.8494C242.5 64.8494 242.873 64.794 243.214 64.6832C243.555 64.5682 243.86 64.3977 244.128 64.1719C244.397 63.9418 244.622 63.6562 244.806 63.3153L246.186 63.7244C245.952 64.2188 245.622 64.6534 245.196 65.0284C244.774 65.3991 244.279 65.6889 243.713 65.8977C243.15 66.1023 242.534 66.2045 241.865 66.2045ZM71.1801 88L73.3534 74.9091H77.7767C78.7994 74.9091 79.607 75.0923 80.1993 75.4588C80.7959 75.8253 81.1986 76.3217 81.4074 76.9482C81.6205 77.5746 81.6631 78.2756 81.5352 79.0511C81.4074 79.8182 81.1325 80.517 80.7107 81.1477C80.2931 81.7784 79.7263 82.2812 79.0104 82.6562C78.2987 83.0312 77.4358 83.2188 76.4216 83.2188H73.2512L73.4813 81.8125H76.6006C77.2994 81.8125 77.8833 81.6911 78.352 81.4482C78.8208 81.2053 79.1872 80.875 79.4514 80.4574C79.7156 80.0398 79.8904 79.571 79.9756 79.0511C80.0608 78.5227 80.0395 78.054 79.9117 77.6449C79.7881 77.2315 79.5303 76.9077 79.1382 76.6733C78.7504 76.4347 78.2029 76.3153 77.4955 76.3153H74.7085L72.7654 88H71.1801ZM89.2361 83.9858L90.2077 78.1818H91.7162L90.0798 88H88.5713L88.8525 86.3381H88.7503C88.4434 86.8366 88.0152 87.2607 87.4655 87.6101C86.9157 87.9553 86.2702 88.1278 85.5287 88.1278C84.915 88.1278 84.3909 87.9936 83.9562 87.7251C83.5258 87.4524 83.219 87.0433 83.0358 86.4979C82.8568 85.9482 82.8355 85.2557 82.9719 84.4205L84.0202 78.1818H85.5287L84.5059 84.3182C84.3909 85.0341 84.4996 85.6051 84.8319 86.0312C85.1643 86.4574 85.6437 86.6705 86.2702 86.6705C86.6452 86.6705 87.0415 86.5746 87.4591 86.3828C87.8809 86.1911 88.2559 85.897 88.5841 85.5007C88.9165 85.1044 89.1338 84.5994 89.2361 83.9858ZM93.0473 88L95.2206 74.9091H96.7291L95.9365 79.7415H96.0644C96.2007 79.571 96.3904 79.3537 96.6333 79.0895C96.8762 78.821 97.2021 78.5824 97.6112 78.3736C98.0203 78.1605 98.5445 78.054 99.1837 78.054C100.01 78.054 100.703 78.2607 101.261 78.674C101.824 79.0874 102.218 79.6733 102.444 80.4318C102.674 81.1903 102.704 82.0852 102.533 83.1165C102.363 84.1562 102.037 85.0575 101.555 85.8203C101.078 86.5788 100.49 87.1669 99.7909 87.5845C99.0963 87.9979 98.3399 88.2045 97.5218 88.2045C96.8911 88.2045 96.4031 88.1001 96.058 87.8913C95.7171 87.6783 95.4656 87.4375 95.3037 87.169C95.146 86.8963 95.0246 86.6705 94.9394 86.4915H94.7604L94.5047 88H93.0473ZM95.3485 83.0909C95.2291 83.8324 95.2313 84.4865 95.3548 85.0533C95.4784 85.6158 95.7235 86.0568 96.0899 86.3764C96.4564 86.6918 96.9422 86.8494 97.5473 86.8494C98.178 86.8494 98.732 86.6832 99.2093 86.3509C99.6865 86.0142 100.079 85.5625 100.385 84.9957C100.692 84.4247 100.905 83.7898 101.025 83.0909C101.135 82.4006 101.131 81.7784 101.012 81.2244C100.897 80.6662 100.654 80.2251 100.283 79.9013C99.9166 79.5732 99.4138 79.4091 98.7746 79.4091C98.161 79.4091 97.6176 79.5646 97.1446 79.8757C96.6759 80.1825 96.2881 80.6129 95.9813 81.1669C95.6787 81.7166 95.4678 82.358 95.3485 83.0909ZM107.704 74.9091L105.531 88H104.022L106.196 74.9091H107.704ZM108.294 88L109.93 78.1818H111.439L109.802 88H108.294ZM111.145 76.5199C110.851 76.5199 110.597 76.4197 110.384 76.2195C110.175 76.0192 110.071 75.7784 110.071 75.4972C110.071 75.2159 110.175 74.9751 110.384 74.7749C110.597 74.5746 110.851 74.4744 111.145 74.4744C111.439 74.4744 111.69 74.5746 111.899 74.7749C112.112 74.9751 112.219 75.2159 112.219 75.4972C112.219 75.7784 112.112 76.0192 111.899 76.2195C111.69 76.4197 111.439 76.5199 111.145 76.5199ZM116.528 88.2045C115.604 88.2045 114.845 87.9872 114.253 87.5526C113.665 87.1179 113.256 86.5192 113.026 85.7564C112.796 84.9936 112.761 84.1222 112.923 83.142C113.085 82.1449 113.413 81.2649 113.908 80.5021C114.402 79.7351 115.018 79.1364 115.755 78.706C116.492 78.2713 117.304 78.054 118.19 78.054C118.881 78.054 119.482 78.1818 119.993 78.4375C120.504 78.6932 120.899 79.0511 121.176 79.5114C121.453 79.9716 121.582 80.5085 121.565 81.1222H120.057C120.014 80.6747 119.823 80.2784 119.482 79.9332C119.145 79.5838 118.646 79.4091 117.986 79.4091C117.398 79.4091 116.857 79.5625 116.362 79.8693C115.868 80.1719 115.453 80.6001 115.116 81.1541C114.779 81.7038 114.551 82.3494 114.432 83.0909C114.304 83.8494 114.315 84.5099 114.464 85.0724C114.613 85.6349 114.884 86.0717 115.276 86.3828C115.668 86.6939 116.162 86.8494 116.759 86.8494C117.151 86.8494 117.517 86.7812 117.858 86.6449C118.199 86.5085 118.502 86.3125 118.766 86.0568C119.03 85.8011 119.239 85.4943 119.392 85.1364H120.901C120.713 85.7159 120.413 86.2379 119.999 86.7024C119.586 87.1626 119.085 87.5291 118.497 87.8018C117.909 88.0703 117.253 88.2045 116.528 88.2045ZM129.677 84.4205L129.958 82.554H130.265L135.302 78.1818H137.168L131.825 82.8097H131.697L129.677 84.4205ZM127.683 88L129.856 74.9091H131.364L129.191 88H127.683ZM133.921 88L130.904 83.142L132.131 82.0938L135.839 88H133.921ZM141.16 88.2045C140.218 88.2045 139.441 87.9957 138.827 87.5781C138.214 87.1562 137.785 86.5682 137.542 85.8139C137.299 85.0554 137.261 84.1733 137.427 83.1676C137.593 82.1619 137.924 81.2756 138.418 80.5085C138.912 79.7372 139.528 79.1364 140.265 78.706C141.003 78.2713 141.812 78.054 142.694 78.054C143.206 78.054 143.698 78.1392 144.171 78.3097C144.644 78.4801 145.053 78.7571 145.398 79.1406C145.743 79.5199 145.984 80.0227 146.12 80.6491C146.257 81.2756 146.248 82.0469 146.095 82.9631L145.993 83.6023H138.424L138.629 82.2983H144.663C144.761 81.7443 144.736 81.25 144.586 80.8153C144.441 80.3807 144.188 80.0376 143.826 79.7862C143.468 79.5348 143.014 79.4091 142.464 79.4091C141.897 79.4091 141.369 79.5625 140.879 79.8693C140.389 80.1719 139.98 80.5554 139.652 81.0199C139.324 81.4801 139.119 81.9489 139.038 82.4261L138.859 83.5511C138.731 84.3651 138.75 85.0128 138.917 85.4943C139.087 85.9759 139.383 86.3232 139.805 86.5362C140.227 86.745 140.755 86.8494 141.39 86.8494C141.795 86.8494 142.168 86.794 142.509 86.6832C142.85 86.5682 143.155 86.3977 143.423 86.1719C143.691 85.9418 143.917 85.6562 144.101 85.3153L145.481 85.7244C145.247 86.2188 144.917 86.6534 144.491 87.0284C144.069 87.3991 143.574 87.6889 143.008 87.8977C142.445 88.1023 141.829 88.2045 141.16 88.2045ZM147.653 91.6818C147.397 91.6818 147.171 91.6605 146.975 91.6179C146.784 91.5795 146.652 91.5412 146.579 91.5028L147.193 90.1733C147.542 90.267 147.853 90.2947 148.126 90.2564C148.403 90.2223 148.665 90.0881 148.912 89.8537C149.164 89.6236 149.426 89.2614 149.698 88.767L150.12 88L148.152 78.1818H149.724L151.13 86.0312H151.233L155.247 78.1818H156.947L150.926 89.419C150.644 89.9432 150.335 90.3714 149.999 90.7038C149.662 91.0362 149.298 91.2813 148.906 91.4389C148.514 91.6009 148.096 91.6818 147.653 91.6818ZM159.243 86.2102L159.013 86.9006C158.86 87.3864 158.664 87.9062 158.425 88.4602C158.187 89.0142 157.95 89.5362 157.716 90.0263C157.481 90.5163 157.292 90.9063 157.147 91.196H155.996C156.09 90.9233 156.218 90.5632 156.38 90.1158C156.542 89.6683 156.714 89.1697 156.897 88.62C157.081 88.0703 157.249 87.5057 157.402 86.9261L157.607 86.2102H159.243ZM175.768 78.2841C175.768 77.581 175.536 77.0483 175.072 76.6861C174.607 76.3196 173.979 76.1364 173.186 76.1364C172.606 76.1364 172.084 76.2301 171.62 76.4176C171.155 76.6051 170.776 76.8629 170.482 77.1911C170.188 77.5192 170.007 77.892 169.939 78.3097C169.883 78.6591 169.92 78.9595 170.047 79.2109C170.175 79.4581 170.356 79.6648 170.591 79.831C170.829 79.9929 171.085 80.1271 171.358 80.2337C171.631 80.3359 171.882 80.419 172.112 80.483L173.391 80.8409C173.71 80.9304 174.066 81.054 174.458 81.2116C174.85 81.3693 175.217 81.5845 175.557 81.8572C175.898 82.1257 176.158 82.4709 176.337 82.8928C176.521 83.3146 176.561 83.8324 176.459 84.446C176.339 85.1534 176.047 85.7926 175.583 86.3636C175.123 86.9347 174.513 87.3885 173.755 87.7251C173.001 88.0618 172.121 88.2301 171.115 88.2301C170.177 88.2301 169.389 88.0788 168.75 87.7763C168.115 87.4737 167.646 87.0518 167.344 86.5107C167.045 85.9695 166.93 85.3409 166.998 84.625H168.635C168.592 85.1193 168.69 85.5284 168.929 85.8523C169.168 86.1719 169.5 86.4105 169.926 86.5682C170.352 86.7216 170.825 86.7983 171.345 86.7983C171.95 86.7983 172.508 86.7003 173.02 86.5043C173.535 86.304 173.962 86.027 174.298 85.6733C174.639 85.3153 174.848 84.8977 174.925 84.4205C175.001 83.9858 174.942 83.6321 174.746 83.3594C174.55 83.0866 174.266 82.8651 173.896 82.6946C173.525 82.5241 173.118 82.375 172.675 82.2472L171.141 81.7869C170.169 81.4929 169.43 81.0732 168.922 80.5277C168.42 79.9822 168.239 79.2685 168.379 78.3864C168.507 77.6534 168.816 77.0142 169.306 76.4688C169.796 75.919 170.403 75.4929 171.128 75.1903C171.852 74.8835 172.632 74.7301 173.467 74.7301C174.311 74.7301 175.033 74.8835 175.634 75.1903C176.235 75.4929 176.682 75.9126 176.976 76.4496C177.271 76.9822 177.379 77.5938 177.302 78.2841H175.768ZM178.272 88L179.909 78.1818H181.417L179.781 88H178.272ZM181.123 76.5199C180.829 76.5199 180.576 76.4197 180.363 76.2195C180.154 76.0192 180.049 75.7784 180.049 75.4972C180.049 75.2159 180.154 74.9751 180.363 74.7749C180.576 74.5746 180.829 74.4744 181.123 74.4744C181.417 74.4744 181.669 74.5746 181.878 74.7749C182.091 74.9751 182.197 75.2159 182.197 75.4972C182.197 75.7784 182.091 76.0192 181.878 76.2195C181.669 76.4197 181.417 76.5199 181.123 76.5199ZM185.893 91.8864C185.16 91.8864 184.547 91.7926 184.052 91.6051C183.558 91.4219 183.162 91.179 182.864 90.8764C182.565 90.5781 182.339 90.2585 182.186 89.9176L183.541 89.0739C183.643 89.2528 183.775 89.4574 183.937 89.6875C184.099 89.9219 184.344 90.1243 184.672 90.2947C185.005 90.4695 185.472 90.5568 186.072 90.5568C186.861 90.5568 187.551 90.3651 188.143 89.9815C188.736 89.598 189.102 88.9972 189.243 88.179L189.55 86.1847H189.422C189.285 86.3636 189.096 86.5852 188.853 86.8494C188.61 87.1094 188.284 87.3416 187.875 87.5462C187.466 87.7464 186.942 87.8466 186.302 87.8466C185.506 87.8466 184.824 87.6591 184.257 87.2841C183.69 86.9091 183.283 86.3636 183.036 85.6477C182.793 84.9318 182.757 84.0625 182.927 83.0398C183.094 82.0341 183.413 81.1584 183.886 80.4126C184.364 79.6626 184.952 79.0831 185.65 78.674C186.349 78.2607 187.112 78.054 187.939 78.054C188.578 78.054 189.068 78.1605 189.409 78.3736C189.754 78.5824 190.006 78.821 190.163 79.0895C190.321 79.3537 190.44 79.571 190.521 79.7415H190.675L190.93 78.1818H192.388L190.726 88.2812C190.585 89.125 190.28 89.8111 189.812 90.3395C189.343 90.8722 188.768 91.2621 188.086 91.5092C187.404 91.7607 186.673 91.8864 185.893 91.8864ZM186.686 86.4915C187.287 86.4915 187.822 86.353 188.29 86.076C188.763 85.799 189.155 85.4006 189.467 84.8807C189.778 84.3608 189.993 83.7386 190.112 83.0142C190.231 82.3068 190.223 81.6825 190.087 81.1413C189.95 80.6001 189.692 80.1761 189.313 79.8693C188.938 79.5625 188.446 79.4091 187.837 79.4091C187.193 79.4091 186.637 79.571 186.168 79.8949C185.704 80.2187 185.327 80.6534 185.037 81.1989C184.747 81.7443 184.547 82.3494 184.436 83.0142C184.325 83.696 184.329 84.299 184.449 84.8232C184.568 85.343 184.811 85.7521 185.177 86.0504C185.548 86.3445 186.051 86.4915 186.686 86.4915ZM195.993 82.0938L195.021 88H193.513L195.149 78.1818H196.606L196.351 79.7159H196.479C196.785 79.2173 197.201 78.8168 197.725 78.5142C198.249 78.2074 198.882 78.054 199.623 78.054C200.28 78.054 200.834 78.1903 201.285 78.4631C201.741 78.7315 202.063 79.1406 202.251 79.6903C202.442 80.2358 202.47 80.9261 202.334 81.7614L201.285 88H199.777L200.8 81.8636C200.927 81.0966 200.827 80.4957 200.499 80.0611C200.175 79.6264 199.662 79.4091 198.959 79.4091C198.473 79.4091 198.023 79.5135 197.61 79.7223C197.201 79.9311 196.854 80.2358 196.568 80.6364C196.287 81.0369 196.095 81.5227 195.993 82.0938ZM206.906 88.2301C206.283 88.2301 205.738 88.1129 205.269 87.8786C204.801 87.6399 204.453 87.2969 204.227 86.8494C204.002 86.3977 203.94 85.8523 204.042 85.2131C204.14 84.6506 204.33 84.1946 204.611 83.8452C204.892 83.4915 205.233 83.2145 205.634 83.0142C206.038 82.8097 206.473 82.6584 206.938 82.5604C207.406 82.4581 207.873 82.3793 208.337 82.3239C208.943 82.2472 209.435 82.1875 209.814 82.1449C210.193 82.1023 210.479 82.032 210.671 81.9339C210.862 81.8359 210.979 81.6676 211.022 81.429V81.3778C211.124 80.7514 211.041 80.2635 210.773 79.9141C210.504 79.5604 210.025 79.3835 209.335 79.3835C208.619 79.3835 208.02 79.5412 207.538 79.8565C207.057 80.1719 206.701 80.5085 206.471 80.8665L205.141 80.3551C205.491 79.7585 205.908 79.294 206.394 78.9616C206.88 78.625 207.389 78.3906 207.922 78.2585C208.455 78.1222 208.968 78.054 209.462 78.054C209.778 78.054 210.134 78.0923 210.53 78.169C210.926 78.2415 211.297 78.3928 211.642 78.6229C211.992 78.853 212.254 79.2003 212.428 79.6648C212.607 80.1293 212.633 80.7514 212.505 81.5312L211.431 88H209.923L210.153 86.6705H210.076C209.94 86.8835 209.731 87.1115 209.45 87.3544C209.173 87.5973 208.821 87.804 208.395 87.9744C207.973 88.1449 207.477 88.2301 206.906 88.2301ZM207.34 86.875C207.937 86.875 208.459 86.7578 208.906 86.5234C209.358 86.2891 209.72 85.9865 209.993 85.6158C210.27 85.245 210.443 84.8551 210.511 84.446L210.741 83.0653C210.664 83.142 210.513 83.2124 210.287 83.2763C210.061 83.3359 209.805 83.3892 209.52 83.4361C209.239 83.4787 208.964 83.517 208.695 83.5511C208.431 83.581 208.218 83.6065 208.056 83.6278C207.656 83.679 207.274 83.7621 206.912 83.8771C206.554 83.9879 206.252 84.1562 206.004 84.3821C205.761 84.6037 205.61 84.9062 205.551 85.2898C205.461 85.8139 205.587 86.2102 205.928 86.4787C206.269 86.7429 206.739 86.875 207.34 86.875ZM220.091 78.1818L219.886 79.4602H214.798L215.003 78.1818H220.091ZM216.869 75.8295H218.378L216.818 85.1875C216.75 85.6136 216.761 85.9332 216.85 86.1463C216.944 86.3551 217.084 86.4957 217.272 86.5682C217.464 86.6364 217.67 86.6705 217.892 86.6705C218.058 86.6705 218.195 86.6619 218.301 86.6449C218.408 86.6236 218.493 86.6065 218.557 86.5938L218.659 87.9489C218.544 87.9872 218.388 88.0256 218.192 88.0639C218.001 88.1065 217.764 88.1278 217.483 88.1278C217.057 88.1278 216.656 88.0362 216.281 87.853C215.906 87.6697 215.619 87.3906 215.418 87.0156C215.218 86.6406 215.165 86.1676 215.258 85.5966L216.869 75.8295ZM227.576 83.9858L228.547 78.1818H230.056L228.42 88H226.911L227.192 86.3381H227.09C226.783 86.8366 226.355 87.2607 225.805 87.6101C225.256 87.9553 224.61 88.1278 223.869 88.1278C223.255 88.1278 222.731 87.9936 222.296 87.7251C221.866 87.4524 221.559 87.0433 221.376 86.4979C221.197 85.9482 221.175 85.2557 221.312 84.4205L222.36 78.1818H223.869L222.846 84.3182C222.731 85.0341 222.839 85.6051 223.172 86.0312C223.504 86.4574 223.984 86.6705 224.61 86.6705C224.985 86.6705 225.381 86.5746 225.799 86.3828C226.221 86.1911 226.596 85.897 226.924 85.5007C227.256 85.1044 227.474 84.5994 227.576 83.9858ZM231.183 88L232.819 78.1818H234.276L234.021 79.6648H234.123C234.379 79.179 234.769 78.7848 235.293 78.4822C235.817 78.1797 236.373 78.0284 236.961 78.0284C237.076 78.0284 237.215 78.0305 237.377 78.0348C237.538 78.0348 237.664 78.0412 237.754 78.054L237.498 79.5881C237.447 79.5753 237.332 79.5561 237.153 79.5305C236.974 79.5007 236.782 79.4858 236.578 79.4858C236.1 79.4858 235.659 79.5859 235.254 79.7862C234.85 79.9822 234.511 80.255 234.238 80.6044C233.965 80.9496 233.791 81.3437 233.714 81.7869L232.691 88H231.183ZM241.707 88.2045C240.765 88.2045 239.988 87.9957 239.374 87.5781C238.76 87.1562 238.332 86.5682 238.089 85.8139C237.846 85.0554 237.808 84.1733 237.974 83.1676C238.14 82.1619 238.471 81.2756 238.965 80.5085C239.459 79.7372 240.075 79.1364 240.812 78.706C241.549 78.2713 242.359 78.054 243.241 78.054C243.753 78.054 244.245 78.1392 244.718 78.3097C245.191 78.4801 245.6 78.7571 245.945 79.1406C246.29 79.5199 246.531 80.0227 246.667 80.6491C246.804 81.2756 246.795 82.0469 246.642 82.9631L246.54 83.6023H238.971L239.176 82.2983H245.21C245.308 81.7443 245.282 81.25 245.133 80.8153C244.988 80.3807 244.735 80.0376 244.373 79.7862C244.015 79.5348 243.561 79.4091 243.011 79.4091C242.444 79.4091 241.916 79.5625 241.426 79.8693C240.936 80.1719 240.527 80.5554 240.199 81.0199C239.87 81.4801 239.666 81.9489 239.585 82.4261L239.406 83.5511C239.278 84.3651 239.297 85.0128 239.464 85.4943C239.634 85.9759 239.93 86.3232 240.352 86.5362C240.774 86.745 241.302 86.8494 241.937 86.8494C242.342 86.8494 242.715 86.794 243.056 86.6832C243.397 86.5682 243.701 86.3977 243.97 86.1719C244.238 85.9418 244.464 85.6562 244.647 85.3153L246.028 85.7244C245.794 86.2188 245.464 86.6534 245.037 87.0284C244.616 87.3991 244.121 87.6889 243.554 87.8977C242.992 88.1023 242.376 88.2045 241.707 88.2045ZM250.685 86.2102L250.455 86.9006C250.301 87.3864 250.105 87.9062 249.867 88.4602C249.628 89.0142 249.391 89.5362 249.157 90.0263C248.923 90.5163 248.733 90.9063 248.588 91.196H247.438C247.531 90.9233 247.659 90.5632 247.821 90.1158C247.983 89.6683 248.156 89.1697 248.339 88.62C248.522 88.0703 248.69 87.5057 248.844 86.9261L249.048 86.2102H250.685ZM259.514 88.1023C259.19 88.1023 258.919 87.9915 258.702 87.7699C258.485 87.544 258.372 87.2713 258.363 86.9517C258.359 86.6321 258.472 86.3658 258.702 86.1527C258.932 85.9354 259.203 85.8267 259.514 85.8267C259.829 85.8267 260.102 85.9375 260.332 86.1591C260.562 86.3807 260.673 86.6449 260.664 86.9517C260.66 87.1648 260.605 87.3587 260.498 87.5334C260.396 87.7081 260.257 87.8466 260.083 87.9489C259.912 88.0511 259.723 88.1023 259.514 88.1023ZM264.471 88.1023C264.147 88.1023 263.876 87.9915 263.659 87.7699C263.442 87.544 263.329 87.2713 263.32 86.9517C263.316 86.6321 263.429 86.3658 263.659 86.1527C263.889 85.9354 264.16 85.8267 264.471 85.8267C264.786 85.8267 265.059 85.9375 265.289 86.1591C265.519 86.3807 265.63 86.6449 265.621 86.9517C265.617 87.1648 265.562 87.3587 265.455 87.5334C265.353 87.7081 265.214 87.8466 265.04 87.9489C264.869 88.0511 264.68 88.1023 264.471 88.1023Z"
+                fill="#F1760D"
+            />
+            <path
+                d="M398.099 232.818V238H396.4V226.364H398.076V230.693H398.184C398.388 230.223 398.701 229.85 399.121 229.574C399.542 229.297 400.091 229.159 400.769 229.159C401.368 229.159 401.89 229.282 402.337 229.528C402.788 229.775 403.137 230.142 403.383 230.631C403.633 231.116 403.758 231.722 403.758 232.449V238H402.059V232.653C402.059 232.013 401.894 231.517 401.565 231.165C401.235 230.809 400.777 230.631 400.19 230.631C399.788 230.631 399.428 230.716 399.11 230.886C398.796 231.057 398.548 231.307 398.366 231.636C398.188 231.962 398.099 232.356 398.099 232.818ZM409.701 238.176C408.883 238.176 408.169 237.989 407.559 237.614C406.949 237.239 406.476 236.714 406.138 236.04C405.801 235.366 405.633 234.578 405.633 233.676C405.633 232.771 405.801 231.979 406.138 231.301C406.476 230.623 406.949 230.097 407.559 229.722C408.169 229.347 408.883 229.159 409.701 229.159C410.519 229.159 411.233 229.347 411.843 229.722C412.453 230.097 412.926 230.623 413.263 231.301C413.601 231.979 413.769 232.771 413.769 233.676C413.769 234.578 413.601 235.366 413.263 236.04C412.926 236.714 412.453 237.239 411.843 237.614C411.233 237.989 410.519 238.176 409.701 238.176ZM409.707 236.75C410.237 236.75 410.676 236.61 411.025 236.33C411.373 236.049 411.631 235.676 411.798 235.21C411.968 234.744 412.053 234.231 412.053 233.67C412.053 233.114 411.968 232.602 411.798 232.136C411.631 231.667 411.373 231.29 411.025 231.006C410.676 230.722 410.237 230.58 409.707 230.58C409.173 230.58 408.729 230.722 408.377 231.006C408.029 231.29 407.769 231.667 407.599 232.136C407.432 232.602 407.349 233.114 407.349 233.67C407.349 234.231 407.432 234.744 407.599 235.21C407.769 235.676 408.029 236.049 408.377 236.33C408.729 236.61 409.173 236.75 409.707 236.75ZM417.292 238L414.724 229.273H416.479L418.19 235.682H418.275L419.991 229.273H421.746L423.451 235.653H423.536L425.235 229.273H426.991L424.428 238H422.695L420.923 231.699H420.792L419.019 238H417.292ZM432.837 238V229.273H434.468V230.693H434.576C434.758 230.212 435.055 229.837 435.468 229.568C435.881 229.295 436.375 229.159 436.951 229.159C437.534 229.159 438.023 229.295 438.417 229.568C438.815 229.841 439.108 230.216 439.298 230.693H439.388C439.597 230.227 439.928 229.856 440.383 229.58C440.837 229.299 441.379 229.159 442.008 229.159C442.799 229.159 443.445 229.407 443.945 229.903C444.449 230.4 444.701 231.148 444.701 232.148V238H443.002V232.307C443.002 231.716 442.841 231.288 442.519 231.023C442.197 230.758 441.813 230.625 441.366 230.625C440.813 230.625 440.383 230.795 440.076 231.136C439.769 231.473 439.616 231.907 439.616 232.438V238H437.923V232.199C437.923 231.725 437.775 231.345 437.479 231.057C437.184 230.769 436.799 230.625 436.326 230.625C436.004 230.625 435.707 230.71 435.434 230.881C435.165 231.047 434.947 231.28 434.781 231.58C434.618 231.879 434.536 232.225 434.536 232.619V238H432.837ZM449.518 238.193C448.965 238.193 448.465 238.091 448.018 237.886C447.571 237.678 447.217 237.377 446.955 236.983C446.698 236.589 446.569 236.106 446.569 235.534C446.569 235.042 446.664 234.636 446.853 234.318C447.042 234 447.298 233.748 447.62 233.562C447.942 233.377 448.302 233.237 448.7 233.142C449.097 233.047 449.503 232.975 449.915 232.926C450.438 232.866 450.862 232.816 451.188 232.778C451.514 232.737 451.751 232.67 451.898 232.58C452.046 232.489 452.12 232.341 452.12 232.136V232.097C452.12 231.6 451.98 231.216 451.7 230.943C451.423 230.67 451.01 230.534 450.461 230.534C449.889 230.534 449.438 230.661 449.109 230.915C448.783 231.165 448.558 231.443 448.433 231.75L446.836 231.386C447.025 230.856 447.302 230.428 447.665 230.102C448.033 229.773 448.455 229.534 448.933 229.386C449.41 229.235 449.912 229.159 450.438 229.159C450.787 229.159 451.156 229.201 451.546 229.284C451.94 229.364 452.308 229.511 452.648 229.727C452.993 229.943 453.275 230.252 453.495 230.653C453.715 231.051 453.825 231.568 453.825 232.205V238H452.165V236.807H452.097C451.987 237.027 451.823 237.242 451.603 237.455C451.383 237.667 451.101 237.843 450.756 237.983C450.412 238.123 449.999 238.193 449.518 238.193ZM449.887 236.83C450.357 236.83 450.758 236.737 451.092 236.551C451.429 236.366 451.684 236.123 451.859 235.824C452.037 235.521 452.126 235.197 452.126 234.852V233.727C452.065 233.788 451.948 233.845 451.773 233.898C451.603 233.947 451.408 233.991 451.188 234.028C450.969 234.062 450.754 234.095 450.546 234.125C450.338 234.152 450.164 234.174 450.023 234.193C449.694 234.235 449.393 234.305 449.12 234.403C448.851 234.502 448.635 234.644 448.472 234.83C448.313 235.011 448.234 235.254 448.234 235.557C448.234 235.977 448.389 236.295 448.7 236.511C449.01 236.723 449.406 236.83 449.887 236.83ZM457.786 232.818V238H456.087V229.273H457.718V230.693H457.826C458.027 230.231 458.341 229.86 458.769 229.58C459.201 229.299 459.745 229.159 460.4 229.159C460.995 229.159 461.515 229.284 461.962 229.534C462.409 229.78 462.756 230.148 463.002 230.636C463.248 231.125 463.371 231.729 463.371 232.449V238H461.673V232.653C461.673 232.021 461.508 231.527 461.178 231.17C460.849 230.811 460.396 230.631 459.82 230.631C459.426 230.631 459.076 230.716 458.769 230.886C458.466 231.057 458.226 231.307 458.048 231.636C457.873 231.962 457.786 232.356 457.786 232.818ZM466.411 241.273C466.157 241.273 465.926 241.252 465.718 241.21C465.51 241.172 465.354 241.131 465.252 241.085L465.661 239.693C465.972 239.777 466.248 239.813 466.491 239.801C466.733 239.79 466.947 239.699 467.133 239.528C467.322 239.358 467.489 239.08 467.633 238.693L467.843 238.114L464.65 229.273H466.468L468.678 236.045H468.769L470.979 229.273H472.803L469.207 239.165C469.04 239.619 468.828 240.004 468.57 240.318C468.313 240.636 468.006 240.875 467.65 241.034C467.294 241.193 466.881 241.273 466.411 241.273ZM478.771 238V226.364H480.469V230.688H480.572C480.67 230.506 480.812 230.295 480.998 230.057C481.183 229.818 481.441 229.61 481.771 229.432C482.1 229.25 482.536 229.159 483.077 229.159C483.782 229.159 484.411 229.337 484.964 229.693C485.517 230.049 485.951 230.562 486.265 231.233C486.583 231.903 486.742 232.71 486.742 233.653C486.742 234.597 486.585 235.405 486.271 236.08C485.956 236.75 485.524 237.267 484.975 237.631C484.426 237.991 483.799 238.17 483.094 238.17C482.564 238.17 482.13 238.081 481.793 237.903C481.46 237.725 481.199 237.517 481.009 237.278C480.82 237.04 480.674 236.828 480.572 236.642H480.43V238H478.771ZM480.435 233.636C480.435 234.25 480.524 234.788 480.702 235.25C480.88 235.712 481.138 236.074 481.475 236.335C481.812 236.593 482.225 236.722 482.714 236.722C483.221 236.722 483.646 236.587 483.987 236.318C484.327 236.045 484.585 235.676 484.759 235.21C484.937 234.744 485.026 234.22 485.026 233.636C485.026 233.061 484.939 232.544 484.765 232.085C484.594 231.627 484.337 231.265 483.992 231C483.651 230.735 483.225 230.602 482.714 230.602C482.221 230.602 481.805 230.729 481.464 230.983C481.127 231.237 480.871 231.591 480.697 232.045C480.522 232.5 480.435 233.03 480.435 233.636ZM489.427 241.273C489.173 241.273 488.942 241.252 488.734 241.21C488.525 241.172 488.37 241.131 488.268 241.085L488.677 239.693C488.987 239.777 489.264 239.813 489.506 239.801C489.749 239.79 489.963 239.699 490.148 239.528C490.338 239.358 490.504 239.08 490.648 238.693L490.859 238.114L487.665 229.273H489.484L491.694 236.045H491.785L493.995 229.273H495.819L492.222 239.165C492.056 239.619 491.844 240.004 491.586 240.318C491.328 240.636 491.022 240.875 490.665 241.034C490.309 241.193 489.897 241.273 489.427 241.273ZM501.489 229.273V230.636H496.722V229.273H501.489ZM498.001 227.182H499.7V235.438C499.7 235.767 499.749 236.015 499.847 236.182C499.946 236.345 500.073 236.456 500.228 236.517C500.387 236.574 500.559 236.602 500.745 236.602C500.881 236.602 501.001 236.593 501.103 236.574C501.205 236.555 501.285 236.54 501.342 236.528L501.648 237.932C501.55 237.97 501.41 238.008 501.228 238.045C501.046 238.087 500.819 238.11 500.546 238.114C500.099 238.121 499.683 238.042 499.296 237.875C498.91 237.708 498.597 237.451 498.359 237.102C498.12 236.754 498.001 236.316 498.001 235.79V227.182ZM507.036 238.176C506.176 238.176 505.436 237.992 504.815 237.625C504.197 237.254 503.72 236.733 503.383 236.062C503.049 235.388 502.883 234.598 502.883 233.693C502.883 232.799 503.049 232.011 503.383 231.33C503.72 230.648 504.19 230.116 504.792 229.733C505.398 229.35 506.106 229.159 506.917 229.159C507.409 229.159 507.887 229.241 508.349 229.403C508.811 229.566 509.226 229.822 509.593 230.17C509.96 230.519 510.25 230.972 510.462 231.528C510.674 232.081 510.781 232.754 510.781 233.545V234.148H503.843V232.875H509.116C509.116 232.428 509.025 232.032 508.843 231.688C508.661 231.339 508.406 231.064 508.076 230.864C507.75 230.663 507.368 230.562 506.928 230.562C506.451 230.562 506.034 230.68 505.678 230.915C505.326 231.146 505.053 231.449 504.86 231.824C504.671 232.195 504.576 232.598 504.576 233.034V234.028C504.576 234.612 504.678 235.108 504.883 235.517C505.091 235.926 505.381 236.239 505.752 236.455C506.123 236.667 506.557 236.773 507.053 236.773C507.375 236.773 507.669 236.727 507.934 236.636C508.199 236.542 508.428 236.402 508.621 236.216C508.815 236.03 508.962 235.801 509.065 235.528L510.673 235.818C510.544 236.292 510.313 236.706 509.979 237.062C509.65 237.415 509.235 237.689 508.735 237.886C508.239 238.08 507.673 238.176 507.036 238.176ZM519.205 231.403L517.665 231.676C517.601 231.479 517.499 231.292 517.359 231.114C517.222 230.936 517.037 230.79 516.802 230.676C516.567 230.562 516.273 230.506 515.921 230.506C515.44 230.506 515.039 230.614 514.717 230.83C514.395 231.042 514.234 231.316 514.234 231.653C514.234 231.945 514.342 232.18 514.558 232.358C514.773 232.536 515.122 232.682 515.603 232.795L516.989 233.114C517.792 233.299 518.391 233.585 518.785 233.972C519.179 234.358 519.376 234.86 519.376 235.477C519.376 236 519.224 236.466 518.921 236.875C518.622 237.28 518.203 237.598 517.665 237.83C517.131 238.061 516.512 238.176 515.808 238.176C514.83 238.176 514.033 237.968 513.415 237.551C512.798 237.131 512.419 236.534 512.279 235.761L513.921 235.511C514.023 235.939 514.234 236.263 514.552 236.483C514.87 236.699 515.285 236.807 515.796 236.807C516.353 236.807 516.798 236.691 517.131 236.46C517.465 236.225 517.631 235.939 517.631 235.602C517.631 235.33 517.529 235.1 517.325 234.915C517.124 234.729 516.815 234.589 516.398 234.494L514.921 234.17C514.107 233.985 513.504 233.689 513.114 233.284C512.728 232.879 512.535 232.366 512.535 231.744C512.535 231.229 512.679 230.778 512.967 230.392C513.254 230.006 513.652 229.705 514.16 229.489C514.667 229.269 515.249 229.159 515.904 229.159C516.847 229.159 517.59 229.364 518.131 229.773C518.673 230.178 519.031 230.722 519.205 231.403ZM527.336 234.653V234.545C527.344 233.841 527.414 233.28 527.546 232.864C527.683 232.447 527.876 232.11 528.126 231.852C528.376 231.595 528.677 231.36 529.029 231.148C529.256 231.004 529.461 230.843 529.643 230.665C529.825 230.487 529.969 230.282 530.075 230.051C530.181 229.82 530.234 229.564 530.234 229.284C530.234 228.947 530.154 228.655 529.995 228.409C529.836 228.163 529.624 227.973 529.359 227.841C529.097 227.705 528.806 227.636 528.484 227.636C528.192 227.636 527.914 227.697 527.648 227.818C527.383 227.939 527.164 228.129 526.989 228.386C526.815 228.64 526.715 228.968 526.688 229.369H524.961C524.987 228.688 525.16 228.112 525.478 227.642C525.796 227.169 526.217 226.811 526.739 226.568C527.266 226.326 527.847 226.205 528.484 226.205C529.181 226.205 529.79 226.335 530.313 226.597C530.836 226.854 531.241 227.216 531.529 227.682C531.821 228.144 531.967 228.684 531.967 229.301C531.967 229.725 531.9 230.108 531.768 230.449C531.635 230.786 531.446 231.087 531.2 231.352C530.957 231.617 530.665 231.852 530.325 232.057C530.003 232.258 529.741 232.466 529.54 232.682C529.344 232.898 529.2 233.153 529.109 233.449C529.018 233.744 528.969 234.11 528.961 234.545V234.653H527.336ZM528.194 238.108C527.883 238.108 527.616 237.998 527.393 237.778C527.169 237.555 527.058 237.286 527.058 236.972C527.058 236.661 527.169 236.396 527.393 236.176C527.616 235.953 527.883 235.841 528.194 235.841C528.501 235.841 528.766 235.953 528.989 236.176C529.217 236.396 529.33 236.661 529.33 236.972C529.33 237.18 529.277 237.371 529.171 237.545C529.069 237.716 528.933 237.852 528.762 237.955C528.592 238.057 528.402 238.108 528.194 238.108Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M914.89 401.364C914.83 400.86 914.588 400.469 914.164 400.19C913.74 399.912 913.22 399.773 912.603 399.773C912.152 399.773 911.758 399.846 911.42 399.991C911.085 400.137 910.823 400.338 910.634 400.593C910.449 400.848 910.356 401.138 910.356 401.463C910.356 401.735 910.421 401.969 910.55 402.164C910.682 402.356 910.851 402.517 911.057 402.646C911.262 402.772 911.478 402.877 911.703 402.96C911.929 403.039 912.136 403.104 912.325 403.153L913.359 403.432C913.624 403.501 913.919 403.598 914.244 403.72C914.572 403.843 914.885 404.01 915.183 404.222C915.485 404.431 915.734 404.7 915.929 405.028C916.125 405.356 916.222 405.759 916.222 406.236C916.222 406.786 916.078 407.283 915.79 407.727C915.505 408.171 915.087 408.524 914.537 408.786C913.99 409.048 913.326 409.179 912.543 409.179C911.814 409.179 911.183 409.061 910.649 408.826C910.119 408.591 909.701 408.263 909.396 407.842C909.095 407.421 908.924 406.932 908.884 406.375H910.157C910.19 406.759 910.319 407.078 910.545 407.33C910.774 407.578 911.062 407.764 911.41 407.886C911.761 408.006 912.139 408.065 912.543 408.065C913.014 408.065 913.437 407.989 913.811 407.837C914.186 407.681 914.482 407.465 914.701 407.19C914.92 406.912 915.029 406.587 915.029 406.216C915.029 405.878 914.935 405.603 914.746 405.391C914.557 405.179 914.308 405.006 914 404.874C913.692 404.741 913.359 404.625 913.001 404.526L911.748 404.168C910.953 403.939 910.323 403.612 909.859 403.188C909.395 402.764 909.163 402.209 909.163 401.523C909.163 400.953 909.317 400.455 909.625 400.031C909.937 399.604 910.354 399.272 910.878 399.037C911.405 398.798 911.993 398.679 912.643 398.679C913.299 398.679 913.882 398.797 914.393 399.032C914.903 399.264 915.308 399.582 915.606 399.987C915.908 400.391 916.067 400.85 916.083 401.364H914.89ZM918.09 409V401.364H919.264V409H918.09ZM918.687 400.091C918.458 400.091 918.261 400.013 918.095 399.857C917.933 399.701 917.852 399.514 917.852 399.295C917.852 399.077 917.933 398.889 918.095 398.734C918.261 398.578 918.458 398.5 918.687 398.5C918.916 398.5 919.111 398.578 919.274 398.734C919.439 398.889 919.522 399.077 919.522 399.295C919.522 399.514 919.439 399.701 919.274 399.857C919.111 400.013 918.916 400.091 918.687 400.091ZM921.194 409V408.105L925.529 402.537V402.457H921.333V401.364H927.041V402.298L922.825 407.827V407.906H927.18V409H921.194ZM931.97 409.159C931.234 409.159 930.599 408.997 930.066 408.672C929.535 408.344 929.126 407.886 928.838 407.3C928.553 406.71 928.41 406.024 928.41 405.241C928.41 404.459 928.553 403.77 928.838 403.173C929.126 402.573 929.527 402.106 930.041 401.771C930.558 401.433 931.161 401.264 931.851 401.264C932.248 401.264 932.641 401.33 933.029 401.463C933.417 401.596 933.77 401.811 934.088 402.109C934.406 402.404 934.66 402.795 934.848 403.283C935.037 403.77 935.132 404.37 935.132 405.082V405.58H929.245V404.565H933.939C933.939 404.134 933.852 403.75 933.68 403.412C933.511 403.074 933.269 402.807 932.954 402.612C932.643 402.416 932.275 402.318 931.851 402.318C931.383 402.318 930.979 402.434 930.638 402.666C930.299 402.895 930.039 403.193 929.857 403.561C929.675 403.929 929.584 404.323 929.584 404.744V405.42C929.584 405.997 929.683 406.486 929.882 406.887C930.084 407.285 930.364 407.588 930.722 407.797C931.08 408.002 931.496 408.105 931.97 408.105C932.278 408.105 932.557 408.062 932.805 407.976C933.057 407.886 933.274 407.754 933.456 407.578C933.639 407.399 933.78 407.177 933.879 406.912L935.013 407.23C934.893 407.615 934.693 407.953 934.411 408.244C934.129 408.533 933.781 408.758 933.367 408.92C932.953 409.08 932.487 409.159 931.97 409.159ZM940.854 409V401.364H942.027V409H940.854ZM941.451 400.091C941.222 400.091 941.025 400.013 940.859 399.857C940.697 399.701 940.616 399.514 940.616 399.295C940.616 399.077 940.697 398.889 940.859 398.734C941.025 398.578 941.222 398.5 941.451 398.5C941.679 398.5 941.875 398.578 942.037 398.734C942.203 398.889 942.286 399.077 942.286 399.295C942.286 399.514 942.203 399.701 942.037 399.857C941.875 400.013 941.679 400.091 941.451 400.091ZM945.35 404.406V409H944.176V401.364H945.31V402.557H945.409C945.588 402.169 945.86 401.857 946.225 401.622C946.589 401.384 947.06 401.264 947.637 401.264C948.154 401.264 948.606 401.37 948.994 401.582C949.382 401.791 949.683 402.109 949.899 402.537C950.114 402.961 950.222 403.498 950.222 404.148V409H949.049V404.227C949.049 403.627 948.893 403.16 948.581 402.825C948.27 402.487 947.842 402.318 947.299 402.318C946.924 402.318 946.589 402.399 946.294 402.562C946.003 402.724 945.772 402.961 945.603 403.273C945.434 403.584 945.35 403.962 945.35 404.406ZM956.462 409V398.818H957.636V402.577H957.735C957.821 402.444 957.941 402.275 958.093 402.07C958.249 401.861 958.471 401.675 958.759 401.513C959.051 401.347 959.445 401.264 959.943 401.264C960.586 401.264 961.152 401.425 961.643 401.746C962.133 402.068 962.516 402.524 962.791 403.114C963.066 403.704 963.204 404.4 963.204 405.202C963.204 406.01 963.066 406.711 962.791 407.305C962.516 407.895 962.135 408.352 961.648 408.677C961.161 408.998 960.599 409.159 959.962 409.159C959.472 409.159 959.079 409.078 958.784 408.915C958.489 408.75 958.262 408.562 958.103 408.354C957.944 408.142 957.821 407.966 957.735 407.827H957.596V409H956.462ZM957.616 405.182C957.616 405.759 957.7 406.267 957.869 406.708C958.038 407.146 958.285 407.489 958.61 407.737C958.935 407.982 959.333 408.105 959.803 408.105C960.294 408.105 960.703 407.976 961.031 407.717C961.363 407.455 961.611 407.104 961.777 406.663C961.946 406.219 962.031 405.725 962.031 405.182C962.031 404.645 961.948 404.161 961.782 403.73C961.62 403.296 961.373 402.953 961.041 402.701C960.713 402.446 960.301 402.318 959.803 402.318C959.326 402.318 958.925 402.439 958.6 402.681C958.275 402.92 958.03 403.254 957.864 403.685C957.699 404.113 957.616 404.612 957.616 405.182ZM965.481 411.864C965.282 411.864 965.105 411.847 964.949 411.814C964.793 411.784 964.685 411.754 964.626 411.724L964.924 410.69C965.209 410.763 965.461 410.79 965.68 410.77C965.899 410.75 966.092 410.652 966.261 410.477C966.434 410.304 966.591 410.024 966.734 409.636L966.953 409.04L964.129 401.364H965.401L967.509 407.449H967.589L969.697 401.364H970.97L967.728 410.114C967.582 410.508 967.402 410.835 967.186 411.093C966.971 411.355 966.72 411.549 966.435 411.675C966.154 411.801 965.836 411.864 965.481 411.864ZM975.839 401.364V402.358H971.882V401.364H975.839ZM973.035 399.534H974.209V406.812C974.209 407.144 974.257 407.393 974.353 407.558C974.452 407.721 974.578 407.83 974.731 407.886C974.886 407.939 975.05 407.966 975.223 407.966C975.352 407.966 975.458 407.959 975.541 407.946C975.624 407.929 975.69 407.916 975.74 407.906L975.978 408.96C975.899 408.99 975.788 409.02 975.645 409.05C975.503 409.083 975.322 409.099 975.103 409.099C974.772 409.099 974.447 409.028 974.129 408.886C973.814 408.743 973.552 408.526 973.343 408.234C973.138 407.943 973.035 407.575 973.035 407.131V399.534ZM980.724 409.159C979.988 409.159 979.353 408.997 978.82 408.672C978.289 408.344 977.88 407.886 977.592 407.3C977.307 406.71 977.164 406.024 977.164 405.241C977.164 404.459 977.307 403.77 977.592 403.173C977.88 402.573 978.281 402.106 978.795 401.771C979.312 401.433 979.915 401.264 980.604 401.264C981.002 401.264 981.395 401.33 981.783 401.463C982.171 401.596 982.524 401.811 982.842 402.109C983.16 402.404 983.413 402.795 983.602 403.283C983.791 403.77 983.886 404.37 983.886 405.082V405.58H977.999V404.565H982.693C982.693 404.134 982.606 403.75 982.434 403.412C982.265 403.074 982.023 402.807 981.708 402.612C981.397 402.416 981.029 402.318 980.604 402.318C980.137 402.318 979.733 402.434 979.391 402.666C979.053 402.895 978.793 403.193 978.611 403.561C978.429 403.929 978.337 404.323 978.337 404.744V405.42C978.337 405.997 978.437 406.486 978.636 406.887C978.838 407.285 979.118 407.588 979.476 407.797C979.834 408.002 980.25 408.105 980.724 408.105C981.032 408.105 981.31 408.062 981.559 407.976C981.811 407.886 982.028 407.754 982.21 407.578C982.393 407.399 982.533 407.177 982.633 406.912L983.766 407.23C983.647 407.615 983.447 407.953 983.165 408.244C982.883 408.533 982.535 408.758 982.121 408.92C981.707 409.08 981.241 409.159 980.724 409.159ZM991.08 403.074L990.026 403.372C989.959 403.196 989.862 403.026 989.732 402.86C989.606 402.691 989.434 402.552 989.215 402.442C988.997 402.333 988.716 402.278 988.375 402.278C987.908 402.278 987.518 402.386 987.207 402.602C986.899 402.814 986.744 403.084 986.744 403.412C986.744 403.704 986.85 403.934 987.063 404.103C987.275 404.272 987.606 404.413 988.057 404.526L989.19 404.804C989.873 404.97 990.382 405.223 990.717 405.565C991.051 405.903 991.219 406.339 991.219 406.872C991.219 407.31 991.093 407.701 990.841 408.045C990.592 408.39 990.244 408.662 989.797 408.861C989.35 409.06 988.829 409.159 988.236 409.159C987.457 409.159 986.812 408.99 986.302 408.652C985.792 408.314 985.468 407.82 985.332 407.17L986.446 406.892C986.552 407.303 986.753 407.611 987.048 407.817C987.346 408.022 987.735 408.125 988.216 408.125C988.763 408.125 989.197 408.009 989.519 407.777C989.843 407.542 990.006 407.26 990.006 406.932C990.006 406.667 989.913 406.445 989.727 406.266C989.542 406.083 989.257 405.947 988.872 405.858L987.6 405.56C986.9 405.394 986.386 405.137 986.058 404.789C985.734 404.438 985.571 403.999 985.571 403.472C985.571 403.041 985.692 402.66 985.934 402.328C986.179 401.997 986.512 401.737 986.933 401.548C987.358 401.359 987.838 401.264 988.375 401.264C989.131 401.264 989.724 401.43 990.155 401.761C990.589 402.093 990.897 402.53 991.08 403.074Z"
+                fill="black"
+            />
+            <path
+                d="M112.072 378.857C111.882 378.857 111.715 378.786 111.572 378.643C111.43 378.5 111.358 378.333 111.358 378.143C111.358 377.952 111.43 377.786 111.572 377.643L113.001 376.214C113.144 376.071 113.311 376 113.501 376C113.692 376 113.858 376.071 114.001 376.214C114.144 376.357 114.215 376.524 114.215 376.714V384.571H114.93C115.12 384.571 115.287 384.643 115.43 384.786C115.572 384.929 115.644 385.095 115.644 385.286C115.644 385.476 115.572 385.643 115.43 385.786C115.287 385.929 115.12 386 114.93 386H112.072C111.882 386 111.715 385.929 111.572 385.786C111.43 385.643 111.358 385.476 111.358 385.286C111.358 385.095 111.43 384.929 111.572 384.786C111.715 384.643 111.882 384.571 112.072 384.571H112.787V378.4L112.572 378.643C112.43 378.786 112.263 378.857 112.072 378.857Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M223.859 386C223.669 386 223.502 385.929 223.359 385.786C223.216 385.643 223.145 385.476 223.145 385.286C223.145 385.095 223.216 384.929 223.359 384.786L228.859 379.257V377.429H223.859C223.669 377.429 223.502 377.357 223.359 377.214C223.216 377.071 223.145 376.905 223.145 376.714C223.145 376.524 223.216 376.357 223.359 376.214C223.502 376.071 223.669 376 223.859 376H229.574C229.764 376 229.931 376.071 230.074 376.214C230.216 376.357 230.288 376.524 230.288 376.714V379.571C230.288 379.762 230.216 379.929 230.074 380.071L224.359 385.786C224.216 385.929 224.05 386 223.859 386ZM232.429 381.714C232.238 381.714 232.071 381.643 231.929 381.5C231.786 381.357 231.714 381.19 231.714 381V376.714C231.714 376.524 231.786 376.357 231.929 376.214C232.071 376.071 232.238 376 232.429 376H238.143C238.333 376 238.5 376.071 238.643 376.214C238.786 376.357 238.857 376.524 238.857 376.714C238.857 376.905 238.786 377.071 238.643 377.214C238.5 377.357 238.333 377.429 238.143 377.429H233.143V380.286H236.714C237.286 380.286 237.786 380.5 238.214 380.929C238.643 381.357 238.857 381.857 238.857 382.429V383.857C238.857 384.429 238.643 384.929 238.214 385.357C237.786 385.786 237.286 386 236.714 386H232.429C232.238 386 232.071 385.929 231.929 385.786C231.786 385.643 231.714 385.476 231.714 385.286C231.714 385.095 231.786 384.929 231.929 384.786C232.071 384.643 232.238 384.571 232.429 384.571H236.714C236.905 384.571 237.071 384.5 237.214 384.357C237.357 384.214 237.429 384.048 237.429 383.857V382.429C237.429 382.238 237.357 382.071 237.214 381.929C237.071 381.786 236.905 381.714 236.714 381.714H232.429Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M372.575 378.857C372.384 378.857 372.217 378.786 372.075 378.643C371.932 378.5 371.86 378.333 371.86 378.143C371.86 377.571 372.075 377.071 372.503 376.643C372.932 376.214 373.432 376 374.003 376H376.86C377.432 376 377.932 376.214 378.36 376.643C378.789 377.071 379.003 377.571 379.003 378.143V379.571C379.003 379.762 378.932 379.929 378.789 380.071L374.26 384.571H378.289C378.479 384.571 378.646 384.643 378.789 384.786C378.932 384.929 379.003 385.095 379.003 385.286C379.003 385.476 378.932 385.643 378.789 385.786C378.646 385.929 378.479 386 378.289 386H372.575C372.384 386 372.217 385.929 372.075 385.786C371.932 385.643 371.86 385.476 371.86 385.286C371.86 385.095 371.932 384.929 372.075 384.786L377.575 379.257V378.143C377.575 377.952 377.503 377.786 377.36 377.643C377.217 377.5 377.051 377.429 376.86 377.429H374.003C373.813 377.429 373.646 377.5 373.503 377.643C373.36 377.786 373.289 377.952 373.289 378.143C373.289 378.333 373.217 378.5 373.075 378.643C372.932 378.786 372.765 378.857 372.575 378.857ZM381.144 381.714C380.953 381.714 380.787 381.643 380.644 381.5C380.501 381.357 380.43 381.19 380.43 381V376.714C380.43 376.524 380.501 376.357 380.644 376.214C380.787 376.071 380.953 376 381.144 376H386.858C387.049 376 387.215 376.071 387.358 376.214C387.501 376.357 387.572 376.524 387.572 376.714C387.572 376.905 387.501 377.071 387.358 377.214C387.215 377.357 387.049 377.429 386.858 377.429H381.858V380.286H385.43C386.001 380.286 386.501 380.5 386.93 380.929C387.358 381.357 387.572 381.857 387.572 382.429V383.857C387.572 384.429 387.358 384.929 386.93 385.357C386.501 385.786 386.001 386 385.43 386H381.144C380.953 386 380.787 385.929 380.644 385.786C380.501 385.643 380.43 385.476 380.43 385.286C380.43 385.095 380.501 384.929 380.644 384.786C380.787 384.643 380.953 384.571 381.144 384.571H385.43C385.62 384.571 385.787 384.5 385.93 384.357C386.072 384.214 386.144 384.048 386.144 383.857V382.429C386.144 382.238 386.072 382.071 385.93 381.929C385.787 381.786 385.62 381.714 385.43 381.714H381.144ZM389.713 381.714C389.523 381.714 389.356 381.643 389.213 381.5C389.07 381.357 388.999 381.19 388.999 381V376.714C388.999 376.524 389.07 376.357 389.213 376.214C389.356 376.071 389.523 376 389.713 376H395.428C395.618 376 395.785 376.071 395.928 376.214C396.07 376.357 396.142 376.524 396.142 376.714C396.142 376.905 396.07 377.071 395.928 377.214C395.785 377.357 395.618 377.429 395.428 377.429H390.428V380.286H393.999C394.57 380.286 395.07 380.5 395.499 380.929C395.928 381.357 396.142 381.857 396.142 382.429V383.857C396.142 384.429 395.928 384.929 395.499 385.357C395.07 385.786 394.57 386 393.999 386H389.713C389.523 386 389.356 385.929 389.213 385.786C389.07 385.643 388.999 385.476 388.999 385.286C388.999 385.095 389.07 384.929 389.213 384.786C389.356 384.643 389.523 384.571 389.713 384.571H393.999C394.189 384.571 394.356 384.5 394.499 384.357C394.642 384.214 394.713 384.048 394.713 383.857V382.429C394.713 382.238 394.642 382.071 394.499 381.929C394.356 381.786 394.189 381.714 393.999 381.714H389.713Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M566.934 386C566.362 386 565.862 385.786 565.434 385.357C565.005 384.929 564.791 384.429 564.791 383.857V378.143C564.791 377.571 565.005 377.071 565.434 376.643C565.862 376.214 566.362 376 566.934 376H569.791C569.981 376 570.148 376.071 570.291 376.214C570.434 376.357 570.505 376.524 570.505 376.714C570.505 376.905 570.434 377.071 570.291 377.214C570.148 377.357 569.981 377.429 569.791 377.429H566.934C566.743 377.429 566.577 377.5 566.434 377.643C566.291 377.786 566.22 377.952 566.22 378.143V380.286H569.791C570.362 380.286 570.862 380.476 571.291 380.857C571.72 381.238 571.934 381.762 571.934 382.429V383.857C571.934 384.429 571.72 384.929 571.291 385.357C570.862 385.786 570.362 386 569.791 386H566.934ZM566.22 381.714V383.857C566.22 384.048 566.291 384.214 566.434 384.357C566.577 384.5 566.743 384.571 566.934 384.571H569.791C569.981 384.571 570.148 384.5 570.291 384.357C570.434 384.214 570.505 384.048 570.505 383.857V382.429C570.505 382.238 570.434 382.071 570.291 381.929C570.148 381.786 569.981 381.714 569.791 381.714H566.22ZM574.075 381.714C573.884 381.714 573.717 381.643 573.575 381.5C573.432 381.357 573.36 381.19 573.36 381V376.714C573.36 376.524 573.432 376.357 573.575 376.214C573.717 376.071 573.884 376 574.075 376H579.789C579.979 376 580.146 376.071 580.289 376.214C580.432 376.357 580.503 376.524 580.503 376.714C580.503 376.905 580.432 377.071 580.289 377.214C580.146 377.357 579.979 377.429 579.789 377.429H574.789V380.286H578.36C578.932 380.286 579.432 380.5 579.86 380.929C580.289 381.357 580.503 381.857 580.503 382.429V383.857C580.503 384.429 580.289 384.929 579.86 385.357C579.432 385.786 578.932 386 578.36 386H574.075C573.884 386 573.717 385.929 573.575 385.786C573.432 385.643 573.36 385.476 573.36 385.286C573.36 385.095 573.432 384.929 573.575 384.786C573.717 384.643 573.884 384.571 574.075 384.571H578.36C578.551 384.571 578.717 384.5 578.86 384.357C579.003 384.214 579.075 384.048 579.075 383.857V382.429C579.075 382.238 579.003 382.071 578.86 381.929C578.717 381.786 578.551 381.714 578.36 381.714H574.075ZM582.644 381.714C582.453 381.714 582.287 381.643 582.144 381.5C582.001 381.357 581.93 381.19 581.93 381V376.714C581.93 376.524 582.001 376.357 582.144 376.214C582.287 376.071 582.453 376 582.644 376H588.358C588.549 376 588.715 376.071 588.858 376.214C589.001 376.357 589.072 376.524 589.072 376.714C589.072 376.905 589.001 377.071 588.858 377.214C588.715 377.357 588.549 377.429 588.358 377.429H583.358V380.286H586.93C587.501 380.286 588.001 380.5 588.43 380.929C588.858 381.357 589.072 381.857 589.072 382.429V383.857C589.072 384.429 588.858 384.929 588.43 385.357C588.001 385.786 587.501 386 586.93 386H582.644C582.453 386 582.287 385.929 582.144 385.786C582.001 385.643 581.93 385.476 581.93 385.286C581.93 385.095 582.001 384.929 582.144 384.786C582.287 384.643 582.453 384.571 582.644 384.571H586.93C587.12 384.571 587.287 384.5 587.43 384.357C587.572 384.214 587.644 384.048 587.644 383.857V382.429C587.644 382.238 587.572 382.071 587.43 381.929C587.287 381.786 587.12 381.714 586.93 381.714H582.644ZM595.499 386H592.642C592.07 386 591.57 385.786 591.142 385.357C590.713 384.929 590.499 384.429 590.499 383.857C590.499 383.667 590.57 383.5 590.713 383.357C590.856 383.214 591.023 383.143 591.213 383.143C591.404 383.143 591.57 383.214 591.713 383.357C591.856 383.5 591.928 383.667 591.928 383.857C591.928 384.048 591.999 384.214 592.142 384.357C592.285 384.5 592.451 384.571 592.642 384.571H595.499C595.689 384.571 595.856 384.5 595.999 384.357C596.142 384.214 596.213 384.048 596.213 383.857V382.429C596.213 382.238 596.142 382.071 595.999 381.929C595.856 381.786 595.689 381.714 595.499 381.714H594.07C593.88 381.714 593.713 381.643 593.57 381.5C593.428 381.357 593.356 381.19 593.356 381C593.356 380.81 593.428 380.643 593.57 380.5C593.713 380.357 593.88 380.286 594.07 380.286H595.499C595.689 380.286 595.856 380.214 595.999 380.071C596.142 379.929 596.213 379.762 596.213 379.571V378.143C596.213 377.952 596.142 377.786 595.999 377.643C595.856 377.5 595.689 377.429 595.499 377.429H592.642C592.451 377.429 592.285 377.5 592.142 377.643C591.999 377.786 591.928 377.952 591.928 378.143C591.928 378.333 591.856 378.5 591.713 378.643C591.57 378.786 591.404 378.857 591.213 378.857C591.023 378.857 590.856 378.786 590.713 378.643C590.57 378.5 590.499 378.333 590.499 378.143C590.499 377.571 590.713 377.071 591.142 376.643C591.57 376.214 592.07 376 592.642 376H595.499C596.07 376 596.57 376.214 596.999 376.643C597.428 377.071 597.642 377.571 597.642 378.143V379.571C597.642 380.143 597.428 380.643 596.999 381.071C597.428 381.405 597.642 381.857 597.642 382.429V383.857C597.642 384.429 597.428 384.929 596.999 385.357C596.57 385.786 596.07 386 595.499 386ZM599.783 381.714C599.592 381.714 599.425 381.643 599.283 381.5C599.14 381.357 599.068 381.19 599.068 381V376.714C599.068 376.524 599.14 376.357 599.283 376.214C599.425 376.071 599.592 376 599.783 376H605.497C605.687 376 605.854 376.071 605.997 376.214C606.14 376.357 606.211 376.524 606.211 376.714C606.211 376.905 606.14 377.071 605.997 377.214C605.854 377.357 605.687 377.429 605.497 377.429H600.497V380.286H604.068C604.64 380.286 605.14 380.5 605.568 380.929C605.997 381.357 606.211 381.857 606.211 382.429V383.857C606.211 384.429 605.997 384.929 605.568 385.357C605.14 385.786 604.64 386 604.068 386H599.783C599.592 386 599.425 385.929 599.283 385.786C599.14 385.643 599.068 385.476 599.068 385.286C599.068 385.095 599.14 384.929 599.283 384.786C599.425 384.643 599.592 384.571 599.783 384.571H604.068C604.259 384.571 604.425 384.5 604.568 384.357C604.711 384.214 604.783 384.048 604.783 383.857V382.429C604.783 382.238 604.711 382.071 604.568 381.929C604.425 381.786 604.259 381.714 604.068 381.714H599.783Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M292.354 217.646C292.158 217.451 291.842 217.451 291.646 217.646L288.464 220.828C288.269 221.024 288.269 221.34 288.464 221.536C288.66 221.731 288.976 221.731 289.172 221.536L292 218.707L294.828 221.536C295.024 221.731 295.34 221.731 295.536 221.536C295.731 221.34 295.731 221.024 295.536 220.828L292.354 217.646ZM292 234H291.5V234.5H292V234ZM291.5 218V234H292.5V218H291.5ZM292 234.5H388V233.5H292V234.5Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M630.646 217.646C630.842 217.451 631.158 217.451 631.354 217.646L634.536 220.828C634.731 221.024 634.731 221.34 634.536 221.536C634.34 221.731 634.024 221.731 633.828 221.536L631 218.707L628.172 221.536C627.976 221.731 627.66 221.731 627.464 221.536C627.269 221.34 627.269 221.024 627.464 220.828L630.646 217.646ZM631 234H631.5V234.5H631V234ZM631.5 218V234H630.5V218H631.5ZM631 234.5H533V233.5H631V234.5Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M898.353 405.35C898.549 405.155 898.549 404.839 898.354 404.643L895.176 401.457C894.981 401.262 894.664 401.261 894.469 401.456C894.273 401.651 894.273 401.968 894.468 402.164L897.293 404.995L894.461 407.82C894.266 408.015 894.265 408.332 894.46 408.527C894.655 408.723 894.972 408.723 895.167 408.528L898.353 405.35ZM88.9994 404.5L897.999 405.496L898.001 404.496L89.0006 403.5L88.9994 404.5Z"
+                fill="#8C8C8C"
+            />
+            <line x1="115.5" y1="412" x2="115.5" y2="397" stroke="#8C8C8C" />
+            <line x1="232.5" y1="412" x2="232.5" y2="397" stroke="#8C8C8C" />
+            <line x1="382.5" y1="412" x2="382.5" y2="397" stroke="#8C8C8C" />
+            <line x1="582.5" y1="412" x2="582.5" y2="397" stroke="#8C8C8C" />
+            <line x1="839.5" y1="413" x2="839.5" y2="398" stroke="#8C8C8C" />
+            <path
+                d="M786.143 381.714H789V376.714C789 376.524 789.071 376.357 789.214 376.214C789.357 376.071 789.524 376 789.714 376C789.905 376 790.071 376.071 790.214 376.214C790.357 376.357 790.429 376.524 790.429 376.714V381.714H791.143C791.333 381.714 791.5 381.786 791.643 381.929C791.786 382.071 791.857 382.238 791.857 382.429C791.857 382.619 791.786 382.786 791.643 382.929C791.5 383.071 791.333 383.143 791.143 383.143H790.429V385.286C790.429 385.476 790.357 385.643 790.214 385.786C790.071 385.929 789.905 386 789.714 386C789.524 386 789.357 385.929 789.214 385.786C789.071 385.643 789 385.476 789 385.286V383.143H785.429C785.238 383.143 785.071 383.071 784.929 382.929C784.786 382.786 784.714 382.619 784.714 382.429V378.143C784.714 377.952 784.786 377.786 784.929 377.643C785.071 377.5 785.238 377.429 785.429 377.429C785.619 377.429 785.786 377.5 785.929 377.643C786.071 377.786 786.143 377.952 786.143 378.143V381.714ZM793.998 378.857C793.807 378.857 793.641 378.786 793.498 378.643C793.355 378.5 793.284 378.333 793.284 378.143C793.284 377.571 793.498 377.071 793.926 376.643C794.355 376.214 794.855 376 795.426 376H798.284C798.855 376 799.355 376.214 799.784 376.643C800.212 377.071 800.426 377.571 800.426 378.143V379.571C800.426 379.762 800.355 379.929 800.212 380.071L795.684 384.571H799.712C799.903 384.571 800.069 384.643 800.212 384.786C800.355 384.929 800.426 385.095 800.426 385.286C800.426 385.476 800.355 385.643 800.212 385.786C800.069 385.929 799.903 386 799.712 386H793.998C793.807 386 793.641 385.929 793.498 385.786C793.355 385.643 793.284 385.476 793.284 385.286C793.284 385.095 793.355 384.929 793.498 384.786L798.998 379.257V378.143C798.998 377.952 798.926 377.786 798.784 377.643C798.641 377.5 798.474 377.429 798.284 377.429H795.426C795.236 377.429 795.069 377.5 794.926 377.643C794.784 377.786 794.712 377.952 794.712 378.143C794.712 378.333 794.641 378.5 794.498 378.643C794.355 378.786 794.188 378.857 793.998 378.857ZM806.853 376C807.424 376 807.924 376.214 808.353 376.643C808.782 377.071 808.996 377.571 808.996 378.143V383.857C808.996 384.429 808.782 384.929 808.353 385.357C807.924 385.786 807.424 386 806.853 386H802.567C802.377 386 802.21 385.929 802.067 385.786C801.924 385.643 801.853 385.476 801.853 385.286C801.853 385.095 801.924 384.929 802.067 384.786C802.21 384.643 802.377 384.571 802.567 384.571H806.853C807.043 384.571 807.21 384.5 807.353 384.357C807.496 384.214 807.567 384.048 807.567 383.857V381.714H803.996C803.424 381.714 802.924 381.524 802.496 381.143C802.067 380.762 801.853 380.238 801.853 379.571V378.143C801.853 377.571 802.067 377.071 802.496 376.643C802.924 376.214 803.424 376 803.996 376H806.853ZM807.567 380.286V378.143C807.567 377.952 807.496 377.786 807.353 377.643C807.21 377.5 807.043 377.429 806.853 377.429H803.996C803.805 377.429 803.639 377.5 803.496 377.643C803.353 377.786 803.282 377.952 803.282 378.143V379.571C803.282 379.762 803.353 379.929 803.496 380.071C803.639 380.214 803.805 380.286 803.996 380.286H807.567ZM811.851 381.714H814.708V376.714C814.708 376.524 814.779 376.357 814.922 376.214C815.065 376.071 815.232 376 815.422 376C815.613 376 815.779 376.071 815.922 376.214C816.065 376.357 816.137 376.524 816.137 376.714V381.714H816.851C817.041 381.714 817.208 381.786 817.351 381.929C817.494 382.071 817.565 382.238 817.565 382.429C817.565 382.619 817.494 382.786 817.351 382.929C817.208 383.071 817.041 383.143 816.851 383.143H816.137V385.286C816.137 385.476 816.065 385.643 815.922 385.786C815.779 385.929 815.613 386 815.422 386C815.232 386 815.065 385.929 814.922 385.786C814.779 385.643 814.708 385.476 814.708 385.286V383.143H811.137C810.946 383.143 810.779 383.071 810.637 382.929C810.494 382.786 810.422 382.619 810.422 382.429V378.143C810.422 377.952 810.494 377.786 810.637 377.643C810.779 377.5 810.946 377.429 811.137 377.429C811.327 377.429 811.494 377.5 811.637 377.643C811.779 377.786 811.851 377.952 811.851 378.143V381.714ZM823.992 376C824.563 376 825.063 376.214 825.492 376.643C825.92 377.071 826.134 377.571 826.134 378.143V383.857C826.134 384.429 825.92 384.929 825.492 385.357C825.063 385.786 824.563 386 823.992 386H819.706C819.515 386 819.349 385.929 819.206 385.786C819.063 385.643 818.992 385.476 818.992 385.286C818.992 385.095 819.063 384.929 819.206 384.786C819.349 384.643 819.515 384.571 819.706 384.571H823.992C824.182 384.571 824.349 384.5 824.492 384.357C824.634 384.214 824.706 384.048 824.706 383.857V381.714H821.134C820.563 381.714 820.063 381.524 819.634 381.143C819.206 380.762 818.992 380.238 818.992 379.571V378.143C818.992 377.571 819.206 377.071 819.634 376.643C820.063 376.214 820.563 376 821.134 376H823.992ZM824.706 380.286V378.143C824.706 377.952 824.634 377.786 824.492 377.643C824.349 377.5 824.182 377.429 823.992 377.429H821.134C820.944 377.429 820.777 377.5 820.634 377.643C820.492 377.786 820.42 377.952 820.42 378.143V379.571C820.42 379.762 820.492 379.929 820.634 380.071C820.777 380.214 820.944 380.286 821.134 380.286H824.706ZM829.704 386C829.132 386 828.632 385.786 828.204 385.357C827.775 384.929 827.561 384.429 827.561 383.857V378.143C827.561 377.571 827.775 377.071 828.204 376.643C828.632 376.214 829.132 376 829.704 376H832.561C832.751 376 832.918 376.071 833.061 376.214C833.204 376.357 833.275 376.524 833.275 376.714C833.275 376.905 833.204 377.071 833.061 377.214C832.918 377.357 832.751 377.429 832.561 377.429H829.704C829.513 377.429 829.347 377.5 829.204 377.643C829.061 377.786 828.99 377.952 828.99 378.143V380.286H832.561C833.132 380.286 833.632 380.476 834.061 380.857C834.49 381.238 834.704 381.762 834.704 382.429V383.857C834.704 384.429 834.49 384.929 834.061 385.357C833.632 385.786 833.132 386 832.561 386H829.704ZM828.99 381.714V383.857C828.99 384.048 829.061 384.214 829.204 384.357C829.347 384.5 829.513 384.571 829.704 384.571H832.561C832.751 384.571 832.918 384.5 833.061 384.357C833.204 384.214 833.275 384.048 833.275 383.857V382.429C833.275 382.238 833.204 382.071 833.061 381.929C832.918 381.786 832.751 381.714 832.561 381.714H828.99ZM836.845 386C836.654 386 836.487 385.929 836.345 385.786C836.202 385.643 836.13 385.476 836.13 385.286C836.13 385.095 836.202 384.929 836.345 384.786L841.845 379.257V377.429H836.845C836.654 377.429 836.487 377.357 836.345 377.214C836.202 377.071 836.13 376.905 836.13 376.714C836.13 376.524 836.202 376.357 836.345 376.214C836.487 376.071 836.654 376 836.845 376H842.559C842.749 376 842.916 376.071 843.059 376.214C843.202 376.357 843.273 376.524 843.273 376.714V379.571C843.273 379.762 843.202 379.929 843.059 380.071L837.345 385.786C837.202 385.929 837.035 386 836.845 386ZM845.414 378.857C845.223 378.857 845.057 378.786 844.914 378.643C844.771 378.5 844.7 378.333 844.7 378.143C844.7 377.571 844.914 377.071 845.342 376.643C845.771 376.214 846.271 376 846.842 376H849.7C850.271 376 850.771 376.214 851.2 376.643C851.628 377.071 851.842 377.571 851.842 378.143V379.571C851.842 379.762 851.771 379.929 851.628 380.071L847.1 384.571H851.128C851.319 384.571 851.485 384.643 851.628 384.786C851.771 384.929 851.842 385.095 851.842 385.286C851.842 385.476 851.771 385.643 851.628 385.786C851.485 385.929 851.319 386 851.128 386H845.414C845.223 386 845.057 385.929 844.914 385.786C844.771 385.643 844.7 385.476 844.7 385.286C844.7 385.095 844.771 384.929 844.914 384.786L850.414 379.257V378.143C850.414 377.952 850.342 377.786 850.2 377.643C850.057 377.5 849.89 377.429 849.7 377.429H846.842C846.652 377.429 846.485 377.5 846.342 377.643C846.2 377.786 846.128 377.952 846.128 378.143C846.128 378.333 846.057 378.5 845.914 378.643C845.771 378.786 845.604 378.857 845.414 378.857ZM858.269 376C858.84 376 859.34 376.214 859.769 376.643C860.198 377.071 860.412 377.571 860.412 378.143V383.857C860.412 384.429 860.198 384.929 859.769 385.357C859.34 385.786 858.84 386 858.269 386H853.983C853.793 386 853.626 385.929 853.483 385.786C853.34 385.643 853.269 385.476 853.269 385.286C853.269 385.095 853.34 384.929 853.483 384.786C853.626 384.643 853.793 384.571 853.983 384.571H858.269C858.459 384.571 858.626 384.5 858.769 384.357C858.912 384.214 858.983 384.048 858.983 383.857V381.714H855.412C854.84 381.714 854.34 381.524 853.912 381.143C853.483 380.762 853.269 380.238 853.269 379.571V378.143C853.269 377.571 853.483 377.071 853.912 376.643C854.34 376.214 854.84 376 855.412 376H858.269ZM858.983 380.286V378.143C858.983 377.952 858.912 377.786 858.769 377.643C858.626 377.5 858.459 377.429 858.269 377.429H855.412C855.221 377.429 855.055 377.5 854.912 377.643C854.769 377.786 854.698 377.952 854.698 378.143V379.571C854.698 379.762 854.769 379.929 854.912 380.071C855.055 380.214 855.221 380.286 855.412 380.286H858.983ZM862.553 381.714C862.362 381.714 862.195 381.643 862.053 381.5C861.91 381.357 861.838 381.19 861.838 381V376.714C861.838 376.524 861.91 376.357 862.053 376.214C862.195 376.071 862.362 376 862.553 376H868.267C868.457 376 868.624 376.071 868.767 376.214C868.91 376.357 868.981 376.524 868.981 376.714C868.981 376.905 868.91 377.071 868.767 377.214C868.624 377.357 868.457 377.429 868.267 377.429H863.267V380.286H866.838C867.41 380.286 867.91 380.5 868.338 380.929C868.767 381.357 868.981 381.857 868.981 382.429V383.857C868.981 384.429 868.767 384.929 868.338 385.357C867.91 385.786 867.41 386 866.838 386H862.553C862.362 386 862.195 385.929 862.053 385.786C861.91 385.643 861.838 385.476 861.838 385.286C861.838 385.095 861.91 384.929 862.053 384.786C862.195 384.643 862.362 384.571 862.553 384.571H866.838C867.029 384.571 867.195 384.5 867.338 384.357C867.481 384.214 867.553 384.048 867.553 383.857V382.429C867.553 382.238 867.481 382.071 867.338 381.929C867.195 381.786 867.029 381.714 866.838 381.714H862.553Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M116 426V435.526C116 438.288 118.239 440.526 121 440.526H162.106C166.061 440.526 169.644 442.857 171.247 446.472L172.368 449L173.085 447.06C174.536 443.133 178.279 440.526 182.465 440.526H230C232.761 440.526 235 438.288 235 435.526V426"
+                stroke="#525252"
+            />
+            <path
+                d="M241 426V436.158C241 438.919 243.239 441.158 246 441.158H297.638C301.405 441.158 304.853 443.275 306.557 446.635L308.263 450L309.419 447.265C310.983 443.563 314.612 441.158 318.63 441.158H378C380.761 441.158 383 438.919 383 436.158V426"
+                stroke="#8C8C8C"
+            />
+            <path
+                d="M389 426V436.158C389 438.919 391.239 441.158 394 441.158H469.531C472.811 441.158 475.883 442.767 477.751 445.463L480.895 450L483.113 446.158C484.9 443.064 488.201 441.158 491.773 441.158H578C580.761 441.158 583 438.919 583 436.158V426"
+                stroke="#B7B7B7"
+            />
+            <path
+                d="M590 426V436.158C590 438.919 592.239 441.158 595 441.158H696.049C698.898 441.158 701.611 442.373 703.508 444.498L708.421 450L712.001 445.189C713.888 442.652 716.863 441.158 720.024 441.158H835C837.761 441.158 840 438.919 840 436.158V426"
+                stroke="#E8E8E8"
+            />
+
+            <motion.rect
+                id="op_push_bytes"
+                x="106.5"
+                y="463.5"
+                width="134"
+                height="33"
+                rx="7.5"
+                fill="#525252"
+                stroke="black"
+                className="animate-shimmer"
+                animate={{ scale: [1, 1.05, 1] }}
+                transition={{
+                    duration: 1,
+                    repeat: Infinity,
+                    repeatType: "mirror"
+                }}
+            />
+            <path
+                d="M126.476 484C125.981 484 125.548 483.814 125.176 483.443C124.805 483.071 124.619 482.638 124.619 482.143V477.19C124.619 476.695 124.805 476.262 125.176 475.89C125.548 475.519 125.981 475.333 126.476 475.333H128.952C129.448 475.333 129.881 475.519 130.252 475.89C130.624 476.262 130.81 476.695 130.81 477.19V482.143C130.81 482.638 130.624 483.071 130.252 483.443C129.881 483.814 129.448 484 128.952 484H126.476ZM129.571 477.19C129.571 477.025 129.51 476.881 129.386 476.757C129.262 476.633 129.117 476.571 128.952 476.571H126.476C126.311 476.571 126.167 476.633 126.043 476.757C125.919 476.881 125.857 477.025 125.857 477.19V482.143C125.857 482.308 125.919 482.452 126.043 482.576C126.167 482.7 126.311 482.762 126.476 482.762H128.952C129.117 482.762 129.262 482.7 129.386 482.576C129.51 482.452 129.571 482.308 129.571 482.143V477.19ZM133.284 479.048H136.379C136.544 479.048 136.689 478.986 136.812 478.862C136.936 478.738 136.998 478.594 136.998 478.429V477.19C136.998 477.025 136.936 476.881 136.812 476.757C136.689 476.633 136.544 476.571 136.379 476.571H133.284V479.048ZM133.284 483.381C133.284 483.546 133.222 483.69 133.098 483.814C132.974 483.938 132.83 484 132.665 484C132.5 484 132.355 483.938 132.232 483.814C132.108 483.69 132.046 483.546 132.046 483.381V475.952C132.046 475.787 132.108 475.643 132.232 475.519C132.355 475.395 132.5 475.333 132.665 475.333H136.379C136.874 475.333 137.308 475.519 137.679 475.89C138.051 476.262 138.236 476.695 138.236 477.19V478.429C138.236 478.924 138.051 479.357 137.679 479.729C137.308 480.1 136.874 480.286 136.379 480.286H133.284V483.381ZM140.092 485.238C139.927 485.238 139.782 485.176 139.658 485.052C139.534 484.929 139.473 484.784 139.473 484.619C139.473 484.454 139.534 484.31 139.658 484.186C139.782 484.062 139.927 484 140.092 484H145.044C145.209 484 145.354 484.062 145.477 484.186C145.601 484.31 145.663 484.454 145.663 484.619C145.663 484.784 145.601 484.929 145.477 485.052C145.354 485.176 145.209 485.238 145.044 485.238H140.092ZM148.137 479.048H151.233C151.398 479.048 151.542 478.986 151.666 478.862C151.79 478.738 151.852 478.594 151.852 478.429V477.19C151.852 477.025 151.79 476.881 151.666 476.757C151.542 476.633 151.398 476.571 151.233 476.571H148.137V479.048ZM148.137 483.381C148.137 483.546 148.076 483.69 147.952 483.814C147.828 483.938 147.683 484 147.518 484C147.353 484 147.209 483.938 147.085 483.814C146.961 483.69 146.899 483.546 146.899 483.381V475.952C146.899 475.787 146.961 475.643 147.085 475.519C147.209 475.395 147.353 475.333 147.518 475.333H151.233C151.728 475.333 152.161 475.519 152.533 475.89C152.904 476.262 153.09 476.695 153.09 477.19V478.429C153.09 478.924 152.904 479.357 152.533 479.729C152.161 480.1 151.728 480.286 151.233 480.286H148.137V483.381ZM158.659 484H156.183C155.688 484 155.255 483.814 154.883 483.443C154.512 483.071 154.326 482.638 154.326 482.143V475.952C154.326 475.787 154.388 475.643 154.512 475.519C154.636 475.395 154.78 475.333 154.945 475.333C155.11 475.333 155.255 475.395 155.378 475.519C155.502 475.643 155.564 475.787 155.564 475.952V482.143C155.564 482.308 155.626 482.452 155.75 482.576C155.874 482.7 156.018 482.762 156.183 482.762H158.659C158.824 482.762 158.969 482.7 159.093 482.576C159.217 482.452 159.278 482.308 159.278 482.143V475.952C159.278 475.787 159.34 475.643 159.464 475.519C159.588 475.395 159.732 475.333 159.898 475.333C160.063 475.333 160.207 475.395 160.331 475.519C160.455 475.643 160.517 475.787 160.517 475.952V482.143C160.517 482.638 160.331 483.071 159.959 483.443C159.588 483.814 159.155 484 158.659 484ZM163.61 480.286C163.032 480.286 162.578 480.121 162.248 479.79C161.918 479.46 161.753 479.006 161.753 478.429V477.19C161.753 476.695 161.918 476.262 162.248 475.89C162.578 475.519 163.032 475.333 163.61 475.333H166.086C166.581 475.333 167.015 475.519 167.386 475.89C167.758 476.262 167.943 476.695 167.943 477.19C167.943 477.356 167.881 477.5 167.758 477.624C167.634 477.748 167.489 477.81 167.324 477.81C167.159 477.81 167.015 477.748 166.891 477.624C166.767 477.5 166.705 477.356 166.705 477.19C166.705 477.025 166.643 476.881 166.52 476.757C166.396 476.633 166.251 476.571 166.086 476.571H163.61C163.445 476.571 163.3 476.633 163.177 476.757C163.053 476.881 162.991 477.025 162.991 477.19V478.429C162.991 478.594 163.053 478.738 163.177 478.862C163.3 478.986 163.445 479.048 163.61 479.048H166.086C166.581 479.048 167.015 479.233 167.386 479.605C167.758 479.976 167.943 480.41 167.943 480.905V482.143C167.943 482.721 167.758 483.175 167.386 483.505C167.015 483.835 166.581 484 166.086 484H163.61C163.115 484 162.681 483.814 162.31 483.443C161.939 483.071 161.753 482.638 161.753 482.143C161.753 481.978 161.815 481.833 161.939 481.71C162.062 481.586 162.207 481.524 162.372 481.524C162.537 481.524 162.681 481.586 162.805 481.71C162.929 481.833 162.991 481.978 162.991 482.143C162.991 482.308 163.053 482.452 163.177 482.576C163.3 482.7 163.445 482.762 163.61 482.762H166.086C166.251 482.762 166.396 482.7 166.52 482.576C166.643 482.452 166.705 482.308 166.705 482.143V480.905C166.705 480.74 166.643 480.595 166.52 480.471C166.396 480.348 166.251 480.286 166.086 480.286H163.61ZM175.37 475.952V483.381C175.37 483.546 175.308 483.69 175.184 483.814C175.061 483.938 174.916 484 174.751 484C174.586 484 174.441 483.938 174.318 483.814C174.194 483.69 174.132 483.546 174.132 483.381V480.286H170.418V483.381C170.418 483.546 170.356 483.69 170.232 483.814C170.108 483.938 169.964 484 169.799 484C169.634 484 169.489 483.938 169.365 483.814C169.241 483.69 169.18 483.546 169.18 483.381V475.952C169.18 475.787 169.241 475.643 169.365 475.519C169.489 475.395 169.634 475.333 169.799 475.333C169.964 475.333 170.108 475.395 170.232 475.519C170.356 475.643 170.418 475.787 170.418 475.952V479.048H174.132V475.952C174.132 475.787 174.194 475.643 174.318 475.519C174.441 475.395 174.586 475.333 174.751 475.333C174.916 475.333 175.061 475.395 175.184 475.519C175.308 475.643 175.37 475.787 175.37 475.952ZM177.225 484C177.06 484 176.916 483.938 176.792 483.814C176.668 483.69 176.606 483.546 176.606 483.381V475.952C176.606 475.787 176.668 475.643 176.792 475.519C176.916 475.395 177.06 475.333 177.225 475.333H180.94C181.435 475.333 181.868 475.519 182.24 475.89C182.611 476.262 182.797 476.695 182.797 477.19V478.429C182.797 478.924 182.611 479.357 182.24 479.729C182.611 480.017 182.797 480.41 182.797 480.905V482.143C182.797 482.638 182.611 483.071 182.24 483.443C181.868 483.814 181.435 484 180.94 484H177.225ZM177.844 479.048H180.94C181.105 479.048 181.249 478.986 181.373 478.862C181.497 478.738 181.559 478.594 181.559 478.429V477.19C181.559 477.025 181.497 476.881 181.373 476.757C181.249 476.633 181.105 476.571 180.94 476.571H177.844V479.048ZM177.844 482.762H180.94C181.105 482.762 181.249 482.7 181.373 482.576C181.497 482.452 181.559 482.308 181.559 482.143V480.905C181.559 480.74 181.497 480.595 181.373 480.471C181.249 480.348 181.105 480.286 180.94 480.286H177.844V482.762ZM186.509 479.914L184.219 477.624C184.095 477.5 184.033 477.356 184.033 477.19V475.952C184.033 475.787 184.095 475.643 184.219 475.519C184.343 475.395 184.487 475.333 184.652 475.333C184.817 475.333 184.962 475.395 185.085 475.519C185.209 475.643 185.271 475.787 185.271 475.952V476.943L187.128 478.8L188.985 476.943V475.952C188.985 475.787 189.047 475.643 189.171 475.519C189.295 475.395 189.439 475.333 189.605 475.333C189.77 475.333 189.914 475.395 190.038 475.519C190.162 475.643 190.224 475.787 190.224 475.952V477.19C190.224 477.356 190.162 477.5 190.038 477.624L187.747 479.89V483.381C187.747 483.546 187.685 483.69 187.562 483.814C187.438 483.938 187.293 484 187.128 484C186.963 484 186.819 483.938 186.695 483.814C186.571 483.69 186.509 483.546 186.509 483.381V479.914ZM192.079 476.571C191.914 476.571 191.769 476.51 191.646 476.386C191.522 476.262 191.46 476.117 191.46 475.952C191.46 475.787 191.522 475.643 191.646 475.519C191.769 475.395 191.914 475.333 192.079 475.333H197.031C197.196 475.333 197.341 475.395 197.465 475.519C197.588 475.643 197.65 475.787 197.65 475.952C197.65 476.117 197.588 476.262 197.465 476.386C197.341 476.51 197.196 476.571 197.031 476.571H195.174V483.381C195.174 483.546 195.112 483.69 194.988 483.814C194.865 483.938 194.72 484 194.555 484C194.39 484 194.246 483.938 194.122 483.814C193.998 483.69 193.936 483.546 193.936 483.381V476.571H192.079ZM203.22 480.286H200.125V482.762H204.458C204.623 482.762 204.768 482.824 204.891 482.948C205.015 483.071 205.077 483.216 205.077 483.381C205.077 483.546 205.015 483.69 204.891 483.814C204.768 483.938 204.623 484 204.458 484H199.506C199.341 484 199.196 483.938 199.072 483.814C198.949 483.69 198.887 483.546 198.887 483.381V475.952C198.887 475.787 198.949 475.643 199.072 475.519C199.196 475.395 199.341 475.333 199.506 475.333H204.458C204.623 475.333 204.768 475.395 204.891 475.519C205.015 475.643 205.077 475.787 205.077 475.952C205.077 476.117 205.015 476.262 204.891 476.386C204.768 476.51 204.623 476.571 204.458 476.571H200.125V479.048H203.22C203.385 479.048 203.529 479.11 203.653 479.233C203.777 479.357 203.839 479.502 203.839 479.667C203.839 479.832 203.777 479.976 203.653 480.1C203.529 480.224 203.385 480.286 203.22 480.286ZM208.171 480.286C207.593 480.286 207.139 480.121 206.809 479.79C206.478 479.46 206.313 479.006 206.313 478.429V477.19C206.313 476.695 206.478 476.262 206.809 475.89C207.139 475.519 207.593 475.333 208.171 475.333H210.647C211.142 475.333 211.575 475.519 211.947 475.89C212.318 476.262 212.504 476.695 212.504 477.19C212.504 477.356 212.442 477.5 212.318 477.624C212.194 477.748 212.05 477.81 211.885 477.81C211.72 477.81 211.575 477.748 211.451 477.624C211.328 477.5 211.266 477.356 211.266 477.19C211.266 477.025 211.204 476.881 211.08 476.757C210.956 476.633 210.812 476.571 210.647 476.571H208.171C208.005 476.571 207.861 476.633 207.737 476.757C207.613 476.881 207.551 477.025 207.551 477.19V478.429C207.551 478.594 207.613 478.738 207.737 478.862C207.861 478.986 208.005 479.048 208.171 479.048H210.647C211.142 479.048 211.575 479.233 211.947 479.605C212.318 479.976 212.504 480.41 212.504 480.905V482.143C212.504 482.721 212.318 483.175 211.947 483.505C211.575 483.835 211.142 484 210.647 484H208.171C207.675 484 207.242 483.814 206.871 483.443C206.499 483.071 206.313 482.638 206.313 482.143C206.313 481.978 206.375 481.833 206.499 481.71C206.623 481.586 206.767 481.524 206.932 481.524C207.098 481.524 207.242 481.586 207.366 481.71C207.49 481.833 207.551 481.978 207.551 482.143C207.551 482.308 207.613 482.452 207.737 482.576C207.861 482.7 208.005 482.762 208.171 482.762H210.647C210.812 482.762 210.956 482.7 211.08 482.576C211.204 482.452 211.266 482.308 211.266 482.143V480.905C211.266 480.74 211.204 480.595 211.08 480.471C210.956 480.348 210.812 480.286 210.647 480.286H208.171ZM214.359 485.238C214.194 485.238 214.05 485.176 213.926 485.052C213.802 484.929 213.74 484.784 213.74 484.619C213.74 484.454 213.802 484.31 213.926 484.186C214.05 484.062 214.194 484 214.359 484H219.312C219.477 484 219.621 484.062 219.745 484.186C219.869 484.31 219.931 484.454 219.931 484.619C219.931 484.784 219.869 484.929 219.745 485.052C219.621 485.176 219.477 485.238 219.312 485.238H214.359ZM222.405 482.39V483.381C222.405 483.546 222.343 483.69 222.219 483.814C222.095 483.938 221.951 484 221.786 484C221.621 484 221.476 483.938 221.353 483.814C221.229 483.69 221.167 483.546 221.167 483.381V482.143C221.167 481.978 221.229 481.833 221.353 481.71L223.395 479.667L221.353 477.624C221.229 477.5 221.167 477.356 221.167 477.19V475.952C221.167 475.787 221.229 475.643 221.353 475.519C221.476 475.395 221.621 475.333 221.786 475.333C221.951 475.333 222.095 475.395 222.219 475.519C222.343 475.643 222.405 475.787 222.405 475.952V476.943L224.262 478.8L226.119 476.943V475.952C226.119 475.787 226.181 475.643 226.305 475.519C226.429 475.395 226.573 475.333 226.738 475.333C226.903 475.333 227.048 475.395 227.172 475.519C227.295 475.643 227.357 475.787 227.357 475.952V477.19C227.357 477.356 227.295 477.5 227.172 477.624L225.104 479.667L227.172 481.71C227.295 481.833 227.357 481.978 227.357 482.143V483.381C227.357 483.546 227.295 483.69 227.172 483.814C227.048 483.938 226.903 484 226.738 484C226.573 484 226.429 483.938 226.305 483.814C226.181 483.69 226.119 483.546 226.119 483.381V482.39L224.262 480.533L222.405 482.39Z"
+                fill="white"
+            />
+
+            <motion.rect
+                id="op_push_data_1"
+                x="248.5"
+                y="463.5"
+                width="134"
+                height="33"
+                rx="7.5"
+                fill="#8C8C8C"
+                stroke="black"
+                className="animate-shimmer"
+                animate={{ scale: [1, 1.05, 1] }}
+                transition={{
+                    duration: 1,
+                    repeat: Infinity,
+                    repeatType: "mirror"
+                }}
+            />
+            <path
+                d="M275.916 484C275.42 484 274.987 483.814 274.616 483.443C274.244 483.071 274.059 482.638 274.059 482.143V477.19C274.059 476.695 274.244 476.262 274.616 475.89C274.987 475.519 275.42 475.333 275.916 475.333H278.392C278.887 475.333 279.32 475.519 279.692 475.89C280.063 476.262 280.249 476.695 280.249 477.19V482.143C280.249 482.638 280.063 483.071 279.692 483.443C279.32 483.814 278.887 484 278.392 484H275.916ZM279.011 477.19C279.011 477.025 278.949 476.881 278.825 476.757C278.701 476.633 278.557 476.571 278.392 476.571H275.916C275.751 476.571 275.606 476.633 275.482 476.757C275.359 476.881 275.297 477.025 275.297 477.19V482.143C275.297 482.308 275.359 482.452 275.482 482.576C275.606 482.7 275.751 482.762 275.916 482.762H278.392C278.557 482.762 278.701 482.7 278.825 482.576C278.949 482.452 279.011 482.308 279.011 482.143V477.19ZM282.723 479.048H285.819C285.984 479.048 286.128 478.986 286.252 478.862C286.376 478.738 286.438 478.594 286.438 478.429V477.19C286.438 477.025 286.376 476.881 286.252 476.757C286.128 476.633 285.984 476.571 285.819 476.571H282.723V479.048ZM282.723 483.381C282.723 483.546 282.661 483.69 282.538 483.814C282.414 483.938 282.269 484 282.104 484C281.939 484 281.795 483.938 281.671 483.814C281.547 483.69 281.485 483.546 281.485 483.381V475.952C281.485 475.787 281.547 475.643 281.671 475.519C281.795 475.395 281.939 475.333 282.104 475.333H285.819C286.314 475.333 286.747 475.519 287.119 475.89C287.49 476.262 287.676 476.695 287.676 477.19V478.429C287.676 478.924 287.49 479.357 287.119 479.729C286.747 480.1 286.314 480.286 285.819 480.286H282.723V483.381ZM289.531 485.238C289.366 485.238 289.222 485.176 289.098 485.052C288.974 484.929 288.912 484.784 288.912 484.619C288.912 484.454 288.974 484.31 289.098 484.186C289.222 484.062 289.366 484 289.531 484H294.483C294.649 484 294.793 484.062 294.917 484.186C295.041 484.31 295.102 484.454 295.102 484.619C295.102 484.784 295.041 484.929 294.917 485.052C294.793 485.176 294.649 485.238 294.483 485.238H289.531ZM297.577 479.048H300.672C300.837 479.048 300.982 478.986 301.105 478.862C301.229 478.738 301.291 478.594 301.291 478.429V477.19C301.291 477.025 301.229 476.881 301.105 476.757C300.982 476.633 300.837 476.571 300.672 476.571H297.577V479.048ZM297.577 483.381C297.577 483.546 297.515 483.69 297.391 483.814C297.267 483.938 297.123 484 296.958 484C296.793 484 296.648 483.938 296.524 483.814C296.401 483.69 296.339 483.546 296.339 483.381V475.952C296.339 475.787 296.401 475.643 296.524 475.519C296.648 475.395 296.793 475.333 296.958 475.333H300.672C301.167 475.333 301.601 475.519 301.972 475.89C302.344 476.262 302.529 476.695 302.529 477.19V478.429C302.529 478.924 302.344 479.357 301.972 479.729C301.601 480.1 301.167 480.286 300.672 480.286H297.577V483.381ZM308.099 484H305.623C305.127 484 304.694 483.814 304.323 483.443C303.951 483.071 303.766 482.638 303.766 482.143V475.952C303.766 475.787 303.827 475.643 303.951 475.519C304.075 475.395 304.22 475.333 304.385 475.333C304.55 475.333 304.694 475.395 304.818 475.519C304.942 475.643 305.004 475.787 305.004 475.952V482.143C305.004 482.308 305.066 482.452 305.189 482.576C305.313 482.7 305.458 482.762 305.623 482.762H308.099C308.264 482.762 308.408 482.7 308.532 482.576C308.656 482.452 308.718 482.308 308.718 482.143V475.952C308.718 475.787 308.78 475.643 308.904 475.519C309.027 475.395 309.172 475.333 309.337 475.333C309.502 475.333 309.646 475.395 309.77 475.519C309.894 475.643 309.956 475.787 309.956 475.952V482.143C309.956 482.638 309.77 483.071 309.399 483.443C309.027 483.814 308.594 484 308.099 484ZM313.049 480.286C312.472 480.286 312.018 480.121 311.688 479.79C311.357 479.46 311.192 479.006 311.192 478.429V477.19C311.192 476.695 311.357 476.262 311.688 475.89C312.018 475.519 312.472 475.333 313.049 475.333H315.526C316.021 475.333 316.454 475.519 316.826 475.89C317.197 476.262 317.383 476.695 317.383 477.19C317.383 477.356 317.321 477.5 317.197 477.624C317.073 477.748 316.929 477.81 316.764 477.81C316.599 477.81 316.454 477.748 316.33 477.624C316.207 477.5 316.145 477.356 316.145 477.19C316.145 477.025 316.083 476.881 315.959 476.757C315.835 476.633 315.691 476.571 315.526 476.571H313.049C312.884 476.571 312.74 476.633 312.616 476.757C312.492 476.881 312.43 477.025 312.43 477.19V478.429C312.43 478.594 312.492 478.738 312.616 478.862C312.74 478.986 312.884 479.048 313.049 479.048H315.526C316.021 479.048 316.454 479.233 316.826 479.605C317.197 479.976 317.383 480.41 317.383 480.905V482.143C317.383 482.721 317.197 483.175 316.826 483.505C316.454 483.835 316.021 484 315.526 484H313.049C312.554 484 312.121 483.814 311.749 483.443C311.378 483.071 311.192 482.638 311.192 482.143C311.192 481.978 311.254 481.833 311.378 481.71C311.502 481.586 311.646 481.524 311.811 481.524C311.976 481.524 312.121 481.586 312.245 481.71C312.368 481.833 312.43 481.978 312.43 482.143C312.43 482.308 312.492 482.452 312.616 482.576C312.74 482.7 312.884 482.762 313.049 482.762H315.526C315.691 482.762 315.835 482.7 315.959 482.576C316.083 482.452 316.145 482.308 316.145 482.143V480.905C316.145 480.74 316.083 480.595 315.959 480.471C315.835 480.348 315.691 480.286 315.526 480.286H313.049ZM324.81 475.952V483.381C324.81 483.546 324.748 483.69 324.624 483.814C324.5 483.938 324.356 484 324.19 484C324.025 484 323.881 483.938 323.757 483.814C323.633 483.69 323.571 483.546 323.571 483.381V480.286H319.857V483.381C319.857 483.546 319.795 483.69 319.671 483.814C319.548 483.938 319.403 484 319.238 484C319.073 484 318.929 483.938 318.805 483.814C318.681 483.69 318.619 483.546 318.619 483.381V475.952C318.619 475.787 318.681 475.643 318.805 475.519C318.929 475.395 319.073 475.333 319.238 475.333C319.403 475.333 319.548 475.395 319.671 475.519C319.795 475.643 319.857 475.787 319.857 475.952V479.048H323.571V475.952C323.571 475.787 323.633 475.643 323.757 475.519C323.881 475.395 324.025 475.333 324.19 475.333C324.356 475.333 324.5 475.395 324.624 475.519C324.748 475.643 324.81 475.787 324.81 475.952ZM326.665 484C326.5 484 326.355 483.938 326.232 483.814C326.108 483.69 326.046 483.546 326.046 483.381V475.952C326.046 475.787 326.108 475.643 326.232 475.519C326.355 475.395 326.5 475.333 326.665 475.333H330.379C330.874 475.333 331.308 475.519 331.679 475.89C332.051 476.262 332.236 476.695 332.236 477.19V482.143C332.236 482.638 332.051 483.071 331.679 483.443C331.308 483.814 330.874 484 330.379 484H326.665ZM330.998 482.143V477.19C330.998 477.025 330.936 476.881 330.812 476.757C330.689 476.633 330.544 476.571 330.379 476.571H327.284V482.762H330.379C330.544 482.762 330.689 482.7 330.812 482.576C330.936 482.452 330.998 482.308 330.998 482.143ZM334.092 484C333.927 484 333.782 483.938 333.658 483.814C333.534 483.69 333.473 483.546 333.473 483.381V477.19C333.473 476.695 333.658 476.262 334.03 475.89C334.401 475.519 334.834 475.333 335.33 475.333H337.806C338.301 475.333 338.734 475.519 339.106 475.89C339.477 476.262 339.663 476.695 339.663 477.19V483.381C339.663 483.546 339.601 483.69 339.477 483.814C339.354 483.938 339.209 484 339.044 484C338.879 484 338.734 483.938 338.611 483.814C338.487 483.69 338.425 483.546 338.425 483.381V480.286H334.711V483.381C334.711 483.546 334.649 483.69 334.525 483.814C334.401 483.938 334.257 484 334.092 484ZM334.711 479.048H338.425V477.19C338.425 477.025 338.363 476.881 338.239 476.757C338.115 476.633 337.971 476.571 337.806 476.571H335.33C335.165 476.571 335.02 476.633 334.896 476.757C334.773 476.881 334.711 477.025 334.711 477.19V479.048ZM341.518 476.571C341.353 476.571 341.209 476.51 341.085 476.386C340.961 476.262 340.899 476.117 340.899 475.952C340.899 475.787 340.961 475.643 341.085 475.519C341.209 475.395 341.353 475.333 341.518 475.333H346.471C346.636 475.333 346.78 475.395 346.904 475.519C347.028 475.643 347.09 475.787 347.09 475.952C347.09 476.117 347.028 476.262 346.904 476.386C346.78 476.51 346.636 476.571 346.471 476.571H344.614V483.381C344.614 483.546 344.552 483.69 344.428 483.814C344.304 483.938 344.16 484 343.995 484C343.829 484 343.685 483.938 343.561 483.814C343.437 483.69 343.376 483.546 343.376 483.381V476.571H341.518ZM348.945 484C348.78 484 348.636 483.938 348.512 483.814C348.388 483.69 348.326 483.546 348.326 483.381V477.19C348.326 476.695 348.512 476.262 348.883 475.89C349.255 475.519 349.688 475.333 350.183 475.333H352.659C353.155 475.333 353.588 475.519 353.959 475.89C354.331 476.262 354.517 476.695 354.517 477.19V483.381C354.517 483.546 354.455 483.69 354.331 483.814C354.207 483.938 354.063 484 353.898 484C353.732 484 353.588 483.938 353.464 483.814C353.34 483.69 353.278 483.546 353.278 483.381V480.286H349.564V483.381C349.564 483.546 349.502 483.69 349.378 483.814C349.255 483.938 349.11 484 348.945 484ZM349.564 479.048H353.278V477.19C353.278 477.025 353.217 476.881 353.093 476.757C352.969 476.633 352.824 476.571 352.659 476.571H350.183C350.018 476.571 349.874 476.633 349.75 476.757C349.626 476.881 349.564 477.025 349.564 477.19V479.048ZM357.61 477.81C357.445 477.81 357.3 477.748 357.177 477.624C357.053 477.5 356.991 477.356 356.991 477.19C356.991 477.025 357.053 476.881 357.177 476.757L358.415 475.519C358.539 475.395 358.683 475.333 358.848 475.333C359.013 475.333 359.158 475.395 359.281 475.519C359.405 475.643 359.467 475.787 359.467 475.952V482.762H360.086C360.251 482.762 360.396 482.824 360.52 482.948C360.643 483.071 360.705 483.216 360.705 483.381C360.705 483.546 360.643 483.69 360.52 483.814C360.396 483.938 360.251 484 360.086 484H357.61C357.445 484 357.3 483.938 357.177 483.814C357.053 483.69 356.991 483.546 356.991 483.381C356.991 483.216 357.053 483.071 357.177 482.948C357.3 482.824 357.445 482.762 357.61 482.762H358.229V477.413L358.043 477.624C357.92 477.748 357.775 477.81 357.61 477.81Z"
+                fill="white"
+            />
+
+            <motion.rect
+                id="op_push_data_2"
+                x="415.5"
+                y="463.5"
+                width="134"
+                height="33"
+                rx="7.5"
+                fill="#B7B7B7"
+                stroke="black"
+                animate={{ scale: [1, 1.05, 1] }}
+                transition={{
+                    duration: 1,
+                    repeat: Infinity,
+                    repeatType: "mirror"
+                }}
+            />
+            <path
+                d="M440.916 484C440.42 484 439.987 483.814 439.616 483.443C439.244 483.071 439.059 482.638 439.059 482.143V477.19C439.059 476.695 439.244 476.262 439.616 475.89C439.987 475.519 440.42 475.333 440.916 475.333H443.392C443.887 475.333 444.32 475.519 444.692 475.89C445.063 476.262 445.249 476.695 445.249 477.19V482.143C445.249 482.638 445.063 483.071 444.692 483.443C444.32 483.814 443.887 484 443.392 484H440.916ZM444.011 477.19C444.011 477.025 443.949 476.881 443.825 476.757C443.701 476.633 443.557 476.571 443.392 476.571H440.916C440.751 476.571 440.606 476.633 440.482 476.757C440.359 476.881 440.297 477.025 440.297 477.19V482.143C440.297 482.308 440.359 482.452 440.482 482.576C440.606 482.7 440.751 482.762 440.916 482.762H443.392C443.557 482.762 443.701 482.7 443.825 482.576C443.949 482.452 444.011 482.308 444.011 482.143V477.19ZM447.723 479.048H450.819C450.984 479.048 451.128 478.986 451.252 478.862C451.376 478.738 451.438 478.594 451.438 478.429V477.19C451.438 477.025 451.376 476.881 451.252 476.757C451.128 476.633 450.984 476.571 450.819 476.571H447.723V479.048ZM447.723 483.381C447.723 483.546 447.661 483.69 447.538 483.814C447.414 483.938 447.269 484 447.104 484C446.939 484 446.795 483.938 446.671 483.814C446.547 483.69 446.485 483.546 446.485 483.381V475.952C446.485 475.787 446.547 475.643 446.671 475.519C446.795 475.395 446.939 475.333 447.104 475.333H450.819C451.314 475.333 451.747 475.519 452.119 475.89C452.49 476.262 452.676 476.695 452.676 477.19V478.429C452.676 478.924 452.49 479.357 452.119 479.729C451.747 480.1 451.314 480.286 450.819 480.286H447.723V483.381ZM454.531 485.238C454.366 485.238 454.222 485.176 454.098 485.052C453.974 484.929 453.912 484.784 453.912 484.619C453.912 484.454 453.974 484.31 454.098 484.186C454.222 484.062 454.366 484 454.531 484H459.483C459.649 484 459.793 484.062 459.917 484.186C460.041 484.31 460.102 484.454 460.102 484.619C460.102 484.784 460.041 484.929 459.917 485.052C459.793 485.176 459.649 485.238 459.483 485.238H454.531ZM462.577 479.048H465.672C465.837 479.048 465.982 478.986 466.105 478.862C466.229 478.738 466.291 478.594 466.291 478.429V477.19C466.291 477.025 466.229 476.881 466.105 476.757C465.982 476.633 465.837 476.571 465.672 476.571H462.577V479.048ZM462.577 483.381C462.577 483.546 462.515 483.69 462.391 483.814C462.267 483.938 462.123 484 461.958 484C461.793 484 461.648 483.938 461.524 483.814C461.401 483.69 461.339 483.546 461.339 483.381V475.952C461.339 475.787 461.401 475.643 461.524 475.519C461.648 475.395 461.793 475.333 461.958 475.333H465.672C466.167 475.333 466.601 475.519 466.972 475.89C467.344 476.262 467.529 476.695 467.529 477.19V478.429C467.529 478.924 467.344 479.357 466.972 479.729C466.601 480.1 466.167 480.286 465.672 480.286H462.577V483.381ZM473.099 484H470.623C470.127 484 469.694 483.814 469.323 483.443C468.951 483.071 468.766 482.638 468.766 482.143V475.952C468.766 475.787 468.827 475.643 468.951 475.519C469.075 475.395 469.22 475.333 469.385 475.333C469.55 475.333 469.694 475.395 469.818 475.519C469.942 475.643 470.004 475.787 470.004 475.952V482.143C470.004 482.308 470.066 482.452 470.189 482.576C470.313 482.7 470.458 482.762 470.623 482.762H473.099C473.264 482.762 473.408 482.7 473.532 482.576C473.656 482.452 473.718 482.308 473.718 482.143V475.952C473.718 475.787 473.78 475.643 473.904 475.519C474.027 475.395 474.172 475.333 474.337 475.333C474.502 475.333 474.646 475.395 474.77 475.519C474.894 475.643 474.956 475.787 474.956 475.952V482.143C474.956 482.638 474.77 483.071 474.399 483.443C474.027 483.814 473.594 484 473.099 484ZM478.049 480.286C477.472 480.286 477.018 480.121 476.688 479.79C476.357 479.46 476.192 479.006 476.192 478.429V477.19C476.192 476.695 476.357 476.262 476.688 475.89C477.018 475.519 477.472 475.333 478.049 475.333H480.526C481.021 475.333 481.454 475.519 481.826 475.89C482.197 476.262 482.383 476.695 482.383 477.19C482.383 477.356 482.321 477.5 482.197 477.624C482.073 477.748 481.929 477.81 481.764 477.81C481.599 477.81 481.454 477.748 481.33 477.624C481.207 477.5 481.145 477.356 481.145 477.19C481.145 477.025 481.083 476.881 480.959 476.757C480.835 476.633 480.691 476.571 480.526 476.571H478.049C477.884 476.571 477.74 476.633 477.616 476.757C477.492 476.881 477.43 477.025 477.43 477.19V478.429C477.43 478.594 477.492 478.738 477.616 478.862C477.74 478.986 477.884 479.048 478.049 479.048H480.526C481.021 479.048 481.454 479.233 481.826 479.605C482.197 479.976 482.383 480.41 482.383 480.905V482.143C482.383 482.721 482.197 483.175 481.826 483.505C481.454 483.835 481.021 484 480.526 484H478.049C477.554 484 477.121 483.814 476.749 483.443C476.378 483.071 476.192 482.638 476.192 482.143C476.192 481.978 476.254 481.833 476.378 481.71C476.502 481.586 476.646 481.524 476.811 481.524C476.976 481.524 477.121 481.586 477.245 481.71C477.368 481.833 477.43 481.978 477.43 482.143C477.43 482.308 477.492 482.452 477.616 482.576C477.74 482.7 477.884 482.762 478.049 482.762H480.526C480.691 482.762 480.835 482.7 480.959 482.576C481.083 482.452 481.145 482.308 481.145 482.143V480.905C481.145 480.74 481.083 480.595 480.959 480.471C480.835 480.348 480.691 480.286 480.526 480.286H478.049ZM489.81 475.952V483.381C489.81 483.546 489.748 483.69 489.624 483.814C489.5 483.938 489.356 484 489.19 484C489.025 484 488.881 483.938 488.757 483.814C488.633 483.69 488.571 483.546 488.571 483.381V480.286H484.857V483.381C484.857 483.546 484.795 483.69 484.671 483.814C484.548 483.938 484.403 484 484.238 484C484.073 484 483.929 483.938 483.805 483.814C483.681 483.69 483.619 483.546 483.619 483.381V475.952C483.619 475.787 483.681 475.643 483.805 475.519C483.929 475.395 484.073 475.333 484.238 475.333C484.403 475.333 484.548 475.395 484.671 475.519C484.795 475.643 484.857 475.787 484.857 475.952V479.048H488.571V475.952C488.571 475.787 488.633 475.643 488.757 475.519C488.881 475.395 489.025 475.333 489.19 475.333C489.356 475.333 489.5 475.395 489.624 475.519C489.748 475.643 489.81 475.787 489.81 475.952ZM491.665 484C491.5 484 491.355 483.938 491.232 483.814C491.108 483.69 491.046 483.546 491.046 483.381V475.952C491.046 475.787 491.108 475.643 491.232 475.519C491.355 475.395 491.5 475.333 491.665 475.333H495.379C495.874 475.333 496.308 475.519 496.679 475.89C497.051 476.262 497.236 476.695 497.236 477.19V482.143C497.236 482.638 497.051 483.071 496.679 483.443C496.308 483.814 495.874 484 495.379 484H491.665ZM495.998 482.143V477.19C495.998 477.025 495.936 476.881 495.812 476.757C495.689 476.633 495.544 476.571 495.379 476.571H492.284V482.762H495.379C495.544 482.762 495.689 482.7 495.812 482.576C495.936 482.452 495.998 482.308 495.998 482.143ZM499.092 484C498.927 484 498.782 483.938 498.658 483.814C498.534 483.69 498.473 483.546 498.473 483.381V477.19C498.473 476.695 498.658 476.262 499.03 475.89C499.401 475.519 499.834 475.333 500.33 475.333H502.806C503.301 475.333 503.734 475.519 504.106 475.89C504.477 476.262 504.663 476.695 504.663 477.19V483.381C504.663 483.546 504.601 483.69 504.477 483.814C504.354 483.938 504.209 484 504.044 484C503.879 484 503.734 483.938 503.611 483.814C503.487 483.69 503.425 483.546 503.425 483.381V480.286H499.711V483.381C499.711 483.546 499.649 483.69 499.525 483.814C499.401 483.938 499.257 484 499.092 484ZM499.711 479.048H503.425V477.19C503.425 477.025 503.363 476.881 503.239 476.757C503.115 476.633 502.971 476.571 502.806 476.571H500.33C500.165 476.571 500.02 476.633 499.896 476.757C499.773 476.881 499.711 477.025 499.711 477.19V479.048ZM506.518 476.571C506.353 476.571 506.209 476.51 506.085 476.386C505.961 476.262 505.899 476.117 505.899 475.952C505.899 475.787 505.961 475.643 506.085 475.519C506.209 475.395 506.353 475.333 506.518 475.333H511.471C511.636 475.333 511.78 475.395 511.904 475.519C512.028 475.643 512.09 475.787 512.09 475.952C512.09 476.117 512.028 476.262 511.904 476.386C511.78 476.51 511.636 476.571 511.471 476.571H509.614V483.381C509.614 483.546 509.552 483.69 509.428 483.814C509.304 483.938 509.16 484 508.995 484C508.829 484 508.685 483.938 508.561 483.814C508.437 483.69 508.376 483.546 508.376 483.381V476.571H506.518ZM513.945 484C513.78 484 513.636 483.938 513.512 483.814C513.388 483.69 513.326 483.546 513.326 483.381V477.19C513.326 476.695 513.512 476.262 513.883 475.89C514.255 475.519 514.688 475.333 515.183 475.333H517.659C518.155 475.333 518.588 475.519 518.959 475.89C519.331 476.262 519.517 476.695 519.517 477.19V483.381C519.517 483.546 519.455 483.69 519.331 483.814C519.207 483.938 519.063 484 518.898 484C518.732 484 518.588 483.938 518.464 483.814C518.34 483.69 518.278 483.546 518.278 483.381V480.286H514.564V483.381C514.564 483.546 514.502 483.69 514.378 483.814C514.255 483.938 514.11 484 513.945 484ZM514.564 479.048H518.278V477.19C518.278 477.025 518.217 476.881 518.093 476.757C517.969 476.633 517.824 476.571 517.659 476.571H515.183C515.018 476.571 514.874 476.633 514.75 476.757C514.626 476.881 514.564 477.025 514.564 477.19V479.048ZM521.372 477.81C521.207 477.81 521.062 477.748 520.939 477.624C520.815 477.5 520.753 477.356 520.753 477.19C520.753 476.695 520.939 476.262 521.31 475.89C521.681 475.519 522.115 475.333 522.61 475.333H525.086C525.581 475.333 526.015 475.519 526.386 475.89C526.758 476.262 526.943 476.695 526.943 477.19V478.429C526.943 478.594 526.881 478.738 526.758 478.862L522.833 482.762H526.324C526.489 482.762 526.634 482.824 526.758 482.948C526.881 483.071 526.943 483.216 526.943 483.381C526.943 483.546 526.881 483.69 526.758 483.814C526.634 483.938 526.489 484 526.324 484H521.372C521.207 484 521.062 483.938 520.939 483.814C520.815 483.69 520.753 483.546 520.753 483.381C520.753 483.216 520.815 483.071 520.939 482.948L525.705 478.156V477.19C525.705 477.025 525.643 476.881 525.52 476.757C525.396 476.633 525.251 476.571 525.086 476.571H522.61C522.445 476.571 522.3 476.633 522.177 476.757C522.053 476.881 521.991 477.025 521.991 477.19C521.991 477.356 521.929 477.5 521.805 477.624C521.681 477.748 521.537 477.81 521.372 477.81Z"
+                fill="black"
+            />
+
+            <motion.rect
+                id="op_push_data_4"
+                x="648.5"
+                y="463.5"
+                width="134"
+                height="33"
+                rx="7.5"
+                fill="#E8E8E8"
+                stroke="black"
+                animate={{ scale: [1, 1.05, 1] }}
+                transition={{
+                    duration: 1,
+                    repeat: Infinity,
+                    repeatType: "mirror"
+                }}
+            />
+            <path
+                d="M673.916 485C673.42 485 672.987 484.814 672.616 484.443C672.244 484.071 672.059 483.638 672.059 483.143V478.19C672.059 477.695 672.244 477.262 672.616 476.89C672.987 476.519 673.42 476.333 673.916 476.333H676.392C676.887 476.333 677.32 476.519 677.692 476.89C678.063 477.262 678.249 477.695 678.249 478.19V483.143C678.249 483.638 678.063 484.071 677.692 484.443C677.32 484.814 676.887 485 676.392 485H673.916ZM677.011 478.19C677.011 478.025 676.949 477.881 676.825 477.757C676.701 477.633 676.557 477.571 676.392 477.571H673.916C673.751 477.571 673.606 477.633 673.482 477.757C673.359 477.881 673.297 478.025 673.297 478.19V483.143C673.297 483.308 673.359 483.452 673.482 483.576C673.606 483.7 673.751 483.762 673.916 483.762H676.392C676.557 483.762 676.701 483.7 676.825 483.576C676.949 483.452 677.011 483.308 677.011 483.143V478.19ZM680.723 480.048H683.819C683.984 480.048 684.128 479.986 684.252 479.862C684.376 479.738 684.438 479.594 684.438 479.429V478.19C684.438 478.025 684.376 477.881 684.252 477.757C684.128 477.633 683.984 477.571 683.819 477.571H680.723V480.048ZM680.723 484.381C680.723 484.546 680.661 484.69 680.538 484.814C680.414 484.938 680.269 485 680.104 485C679.939 485 679.795 484.938 679.671 484.814C679.547 484.69 679.485 484.546 679.485 484.381V476.952C679.485 476.787 679.547 476.643 679.671 476.519C679.795 476.395 679.939 476.333 680.104 476.333H683.819C684.314 476.333 684.747 476.519 685.119 476.89C685.49 477.262 685.676 477.695 685.676 478.19V479.429C685.676 479.924 685.49 480.357 685.119 480.729C684.747 481.1 684.314 481.286 683.819 481.286H680.723V484.381ZM687.531 486.238C687.366 486.238 687.222 486.176 687.098 486.052C686.974 485.929 686.912 485.784 686.912 485.619C686.912 485.454 686.974 485.31 687.098 485.186C687.222 485.062 687.366 485 687.531 485H692.483C692.649 485 692.793 485.062 692.917 485.186C693.041 485.31 693.102 485.454 693.102 485.619C693.102 485.784 693.041 485.929 692.917 486.052C692.793 486.176 692.649 486.238 692.483 486.238H687.531ZM695.577 480.048H698.672C698.837 480.048 698.982 479.986 699.105 479.862C699.229 479.738 699.291 479.594 699.291 479.429V478.19C699.291 478.025 699.229 477.881 699.105 477.757C698.982 477.633 698.837 477.571 698.672 477.571H695.577V480.048ZM695.577 484.381C695.577 484.546 695.515 484.69 695.391 484.814C695.267 484.938 695.123 485 694.958 485C694.793 485 694.648 484.938 694.524 484.814C694.401 484.69 694.339 484.546 694.339 484.381V476.952C694.339 476.787 694.401 476.643 694.524 476.519C694.648 476.395 694.793 476.333 694.958 476.333H698.672C699.167 476.333 699.601 476.519 699.972 476.89C700.344 477.262 700.529 477.695 700.529 478.19V479.429C700.529 479.924 700.344 480.357 699.972 480.729C699.601 481.1 699.167 481.286 698.672 481.286H695.577V484.381ZM706.099 485H703.623C703.127 485 702.694 484.814 702.323 484.443C701.951 484.071 701.766 483.638 701.766 483.143V476.952C701.766 476.787 701.827 476.643 701.951 476.519C702.075 476.395 702.22 476.333 702.385 476.333C702.55 476.333 702.694 476.395 702.818 476.519C702.942 476.643 703.004 476.787 703.004 476.952V483.143C703.004 483.308 703.066 483.452 703.189 483.576C703.313 483.7 703.458 483.762 703.623 483.762H706.099C706.264 483.762 706.408 483.7 706.532 483.576C706.656 483.452 706.718 483.308 706.718 483.143V476.952C706.718 476.787 706.78 476.643 706.904 476.519C707.027 476.395 707.172 476.333 707.337 476.333C707.502 476.333 707.646 476.395 707.77 476.519C707.894 476.643 707.956 476.787 707.956 476.952V483.143C707.956 483.638 707.77 484.071 707.399 484.443C707.027 484.814 706.594 485 706.099 485ZM711.049 481.286C710.472 481.286 710.018 481.121 709.688 480.79C709.357 480.46 709.192 480.006 709.192 479.429V478.19C709.192 477.695 709.357 477.262 709.688 476.89C710.018 476.519 710.472 476.333 711.049 476.333H713.526C714.021 476.333 714.454 476.519 714.826 476.89C715.197 477.262 715.383 477.695 715.383 478.19C715.383 478.356 715.321 478.5 715.197 478.624C715.073 478.748 714.929 478.81 714.764 478.81C714.599 478.81 714.454 478.748 714.33 478.624C714.207 478.5 714.145 478.356 714.145 478.19C714.145 478.025 714.083 477.881 713.959 477.757C713.835 477.633 713.691 477.571 713.526 477.571H711.049C710.884 477.571 710.74 477.633 710.616 477.757C710.492 477.881 710.43 478.025 710.43 478.19V479.429C710.43 479.594 710.492 479.738 710.616 479.862C710.74 479.986 710.884 480.048 711.049 480.048H713.526C714.021 480.048 714.454 480.233 714.826 480.605C715.197 480.976 715.383 481.41 715.383 481.905V483.143C715.383 483.721 715.197 484.175 714.826 484.505C714.454 484.835 714.021 485 713.526 485H711.049C710.554 485 710.121 484.814 709.749 484.443C709.378 484.071 709.192 483.638 709.192 483.143C709.192 482.978 709.254 482.833 709.378 482.71C709.502 482.586 709.646 482.524 709.811 482.524C709.976 482.524 710.121 482.586 710.245 482.71C710.368 482.833 710.43 482.978 710.43 483.143C710.43 483.308 710.492 483.452 710.616 483.576C710.74 483.7 710.884 483.762 711.049 483.762H713.526C713.691 483.762 713.835 483.7 713.959 483.576C714.083 483.452 714.145 483.308 714.145 483.143V481.905C714.145 481.74 714.083 481.595 713.959 481.471C713.835 481.348 713.691 481.286 713.526 481.286H711.049ZM722.81 476.952V484.381C722.81 484.546 722.748 484.69 722.624 484.814C722.5 484.938 722.356 485 722.19 485C722.025 485 721.881 484.938 721.757 484.814C721.633 484.69 721.571 484.546 721.571 484.381V481.286H717.857V484.381C717.857 484.546 717.795 484.69 717.671 484.814C717.548 484.938 717.403 485 717.238 485C717.073 485 716.929 484.938 716.805 484.814C716.681 484.69 716.619 484.546 716.619 484.381V476.952C716.619 476.787 716.681 476.643 716.805 476.519C716.929 476.395 717.073 476.333 717.238 476.333C717.403 476.333 717.548 476.395 717.671 476.519C717.795 476.643 717.857 476.787 717.857 476.952V480.048H721.571V476.952C721.571 476.787 721.633 476.643 721.757 476.519C721.881 476.395 722.025 476.333 722.19 476.333C722.356 476.333 722.5 476.395 722.624 476.519C722.748 476.643 722.81 476.787 722.81 476.952ZM724.665 485C724.5 485 724.355 484.938 724.232 484.814C724.108 484.69 724.046 484.546 724.046 484.381V476.952C724.046 476.787 724.108 476.643 724.232 476.519C724.355 476.395 724.5 476.333 724.665 476.333H728.379C728.874 476.333 729.308 476.519 729.679 476.89C730.051 477.262 730.236 477.695 730.236 478.19V483.143C730.236 483.638 730.051 484.071 729.679 484.443C729.308 484.814 728.874 485 728.379 485H724.665ZM728.998 483.143V478.19C728.998 478.025 728.936 477.881 728.812 477.757C728.689 477.633 728.544 477.571 728.379 477.571H725.284V483.762H728.379C728.544 483.762 728.689 483.7 728.812 483.576C728.936 483.452 728.998 483.308 728.998 483.143ZM732.092 485C731.927 485 731.782 484.938 731.658 484.814C731.534 484.69 731.473 484.546 731.473 484.381V478.19C731.473 477.695 731.658 477.262 732.03 476.89C732.401 476.519 732.834 476.333 733.33 476.333H735.806C736.301 476.333 736.734 476.519 737.106 476.89C737.477 477.262 737.663 477.695 737.663 478.19V484.381C737.663 484.546 737.601 484.69 737.477 484.814C737.354 484.938 737.209 485 737.044 485C736.879 485 736.734 484.938 736.611 484.814C736.487 484.69 736.425 484.546 736.425 484.381V481.286H732.711V484.381C732.711 484.546 732.649 484.69 732.525 484.814C732.401 484.938 732.257 485 732.092 485ZM732.711 480.048H736.425V478.19C736.425 478.025 736.363 477.881 736.239 477.757C736.115 477.633 735.971 477.571 735.806 477.571H733.33C733.165 477.571 733.02 477.633 732.896 477.757C732.773 477.881 732.711 478.025 732.711 478.19V480.048ZM739.518 477.571C739.353 477.571 739.209 477.51 739.085 477.386C738.961 477.262 738.899 477.117 738.899 476.952C738.899 476.787 738.961 476.643 739.085 476.519C739.209 476.395 739.353 476.333 739.518 476.333H744.471C744.636 476.333 744.78 476.395 744.904 476.519C745.028 476.643 745.09 476.787 745.09 476.952C745.09 477.117 745.028 477.262 744.904 477.386C744.78 477.51 744.636 477.571 744.471 477.571H742.614V484.381C742.614 484.546 742.552 484.69 742.428 484.814C742.304 484.938 742.16 485 741.995 485C741.829 485 741.685 484.938 741.561 484.814C741.437 484.69 741.376 484.546 741.376 484.381V477.571H739.518ZM746.945 485C746.78 485 746.636 484.938 746.512 484.814C746.388 484.69 746.326 484.546 746.326 484.381V478.19C746.326 477.695 746.512 477.262 746.883 476.89C747.255 476.519 747.688 476.333 748.183 476.333H750.659C751.155 476.333 751.588 476.519 751.959 476.89C752.331 477.262 752.517 477.695 752.517 478.19V484.381C752.517 484.546 752.455 484.69 752.331 484.814C752.207 484.938 752.063 485 751.898 485C751.732 485 751.588 484.938 751.464 484.814C751.34 484.69 751.278 484.546 751.278 484.381V481.286H747.564V484.381C747.564 484.546 747.502 484.69 747.378 484.814C747.255 484.938 747.11 485 746.945 485ZM747.564 480.048H751.278V478.19C751.278 478.025 751.217 477.881 751.093 477.757C750.969 477.633 750.824 477.571 750.659 477.571H748.183C748.018 477.571 747.874 477.633 747.75 477.757C747.626 477.881 747.564 478.025 747.564 478.19V480.048ZM754.991 481.286H757.467V476.952C757.467 476.787 757.529 476.643 757.653 476.519C757.777 476.395 757.921 476.333 758.086 476.333C758.251 476.333 758.396 476.395 758.52 476.519C758.643 476.643 758.705 476.787 758.705 476.952V481.286H759.324C759.489 481.286 759.634 481.348 759.758 481.471C759.881 481.595 759.943 481.74 759.943 481.905C759.943 482.07 759.881 482.214 759.758 482.338C759.634 482.462 759.489 482.524 759.324 482.524H758.705V484.381C758.705 484.546 758.643 484.69 758.52 484.814C758.396 484.938 758.251 485 758.086 485C757.921 485 757.777 484.938 757.653 484.814C757.529 484.69 757.467 484.546 757.467 484.381V482.524H754.372C754.207 482.524 754.062 482.462 753.939 482.338C753.815 482.214 753.753 482.07 753.753 481.905V478.19C753.753 478.025 753.815 477.881 753.939 477.757C754.062 477.633 754.207 477.571 754.372 477.571C754.537 477.571 754.681 477.633 754.805 477.757C754.929 477.881 754.991 478.025 754.991 478.19V481.286Z"
+                fill="black"
+            />
+
+            <motion.line
+                x1="232"
+                y1="404.5"
+                x2="89"
+                y2="404.5"
+                stroke="#525252"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="382"
+                y1="404.5"
+                x2="233"
+                y2="404.5"
+                stroke="#8C8C8C"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="382"
+                y1="404.5"
+                x2="233"
+                y2="404.5"
+                stroke="#8C8C8C"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="581.999"
+                y1="404.5"
+                x2="382.999"
+                y2="404.374"
+                stroke="#8C8C8C"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="581.999"
+                y1="404.5"
+                x2="382.999"
+                y2="404.374"
+                stroke="#B7B7B7"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="839"
+                y1="404.5"
+                x2="583"
+                y2="404.5"
+                stroke="#8C8C8C"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="839"
+                y1="404.5"
+                x2="583"
+                y2="404.5"
+                stroke="#E8E8E8"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="839"
+                y1="404.5"
+                x2="583"
+                y2="404.5"
+                stroke="#8C8C8C"
+                strokeWidth="3"
+            />
+            <motion.line
+                x1="839"
+                y1="404.5"
+                x2="583"
+                y2="404.5"
+                stroke="#E8E8E8"
+                strokeWidth="3"
+            />
+            <path
+                d="M284.354 174.354C284.549 174.158 284.549 173.842 284.354 173.646L281.172 170.464C280.976 170.269 280.66 170.269 280.464 170.464C280.269 170.66 280.269 170.976 280.464 171.172L283.293 174L280.464 176.828C280.269 177.024 280.269 177.34 280.464 177.536C280.66 177.731 280.976 177.731 281.172 177.536L284.354 174.354ZM170.005 103.928L169.942 104.364L170.931 104.508L170.995 104.072L170.005 103.928ZM231.643 155.363L232.767 157.616L233.662 157.17L232.538 154.916L231.643 155.363ZM260.059 174.5H284V173.5H260.059V174.5ZM232.767 157.616C237.929 167.962 248.497 174.5 260.059 174.5V173.5C248.876 173.5 238.655 167.177 233.662 157.17L232.767 157.616ZM202.735 139.139C214.78 138.118 226.247 144.546 231.643 155.363L232.538 154.916C226.958 143.733 215.103 137.087 202.651 138.143L202.735 139.139ZM169.942 104.364C167.116 123.802 183.165 140.799 202.735 139.139L202.651 138.143C183.721 139.748 168.199 123.307 170.931 104.508L169.942 104.364Z"
+                fill="#E77429"
+            />
+            <path
+                d="M477.588 318.628V271H468.029V318.628H460L472.5 335L485.5 318.628H477.588Z"
+                stroke="#8C8C8C"
+            />
+            <path
+                d="M292 126L295.5 128.021L295.5 123.979L292 126ZM421 125.65L418.068 125.65L418.068 126.35L421 126.35L421 125.65ZM412.205 125.65L406.341 125.65L406.341 126.35L412.205 126.35L412.205 125.65ZM400.477 125.65L394.614 125.65L394.614 126.35L400.477 126.35L400.477 125.65ZM388.75 125.65L382.886 125.65L382.886 126.35L388.75 126.35L388.75 125.65ZM377.023 125.65L371.159 125.65L371.159 126.35L377.023 126.35L377.023 125.65ZM365.295 125.65L359.432 125.65L359.432 126.35L365.295 126.35L365.295 125.65ZM353.568 125.65L347.705 125.65L347.705 126.35L353.568 126.35L353.568 125.65ZM341.841 125.65L335.977 125.65L335.977 126.35L341.841 126.35L341.841 125.65ZM330.114 125.65L324.25 125.65L324.25 126.35L330.114 126.35L330.114 125.65ZM318.386 125.65L312.523 125.65L312.523 126.35L318.386 126.35L318.386 125.65ZM306.659 125.65L300.795 125.65L300.795 126.35L306.659 126.35L306.659 125.65ZM292 126L295.5 128.021L295.5 123.979L292 126ZM421 125.65L418.068 125.65L418.068 126.35L421 126.35L421 125.65ZM412.205 125.65L406.341 125.65L406.341 126.35L412.205 126.35L412.205 125.65ZM400.477 125.65L394.614 125.65L394.614 126.35L400.477 126.35L400.477 125.65ZM388.75 125.65L382.886 125.65L382.886 126.35L388.75 126.35L388.75 125.65ZM377.023 125.65L371.159 125.65L371.159 126.35L377.023 126.35L377.023 125.65ZM365.295 125.65L359.432 125.65L359.432 126.35L365.295 126.35L365.295 125.65ZM353.568 125.65L347.705 125.65L347.705 126.35L353.568 126.35L353.568 125.65ZM341.841 125.65L335.977 125.65L335.977 126.35L341.841 126.35L341.841 125.65ZM330.114 125.65L324.25 125.65L324.25 126.35L330.114 126.35L330.114 125.65ZM318.386 125.65L312.523 125.65L312.523 126.35L318.386 126.35L318.386 125.65ZM306.659 125.65L300.795 125.65L300.795 126.35L306.659 126.35L306.659 125.65Z"
+                fill="#8C8C8C"
+            />
+            <path
+                d="M631 126L627.5 123.979V128.021L631 126ZM501 126.35H503.955V125.65H501V126.35ZM509.864 126.35H515.773V125.65H509.864V126.35ZM521.682 126.35H527.591V125.65H521.682V126.35ZM533.5 126.35H539.409V125.65H533.5V126.35ZM545.318 126.35H551.227V125.65H545.318V126.35ZM557.136 126.35H563.045V125.65H557.136V126.35ZM568.955 126.35H574.864V125.65H568.955V126.35ZM580.773 126.35H586.682V125.65H580.773V126.35ZM592.591 126.35H598.5V125.65H592.591V126.35ZM604.409 126.35H610.318V125.65H604.409V126.35ZM616.227 126.35H622.136V125.65H616.227V126.35ZM631 126L627.5 123.979V128.021L631 126ZM501 126.35H503.955V125.65H501V126.35ZM509.864 126.35H515.773V125.65H509.864V126.35ZM521.682 126.35H527.591V125.65H521.682V126.35ZM533.5 126.35H539.409V125.65H533.5V126.35ZM545.318 126.35H551.227V125.65H545.318V126.35ZM557.136 126.35H563.045V125.65H557.136V126.35ZM568.955 126.35H574.864V125.65H568.955V126.35ZM580.773 126.35H586.682V125.65H580.773V126.35ZM592.591 126.35H598.5V125.65H592.591V126.35ZM604.409 126.35H610.318V125.65H604.409V126.35ZM616.227 126.35H622.136V125.65H616.227V126.35Z"
+                fill="#8C8C8C"
+            />
+        </svg>
+    )
+}
